Split the child-modify reconciliation code out into its own file. Split the row- and column-store reconciliation code out into their own files.
