+++ src/reconcile/rec_write.c
@@ -8,38 +8,13 @@
 
 #include "wt_internal.h"
 
-static void __rec_cell_build_addr(
-		WT_SESSION_IMPL *, WT_RECONCILE *, WT_ADDR *, bool, uint64_t);
-static int  __rec_cell_build_int_key(WT_SESSION_IMPL *,
-		WT_RECONCILE *, const void *, size_t, bool *);
-static int  __rec_cell_build_leaf_key(WT_SESSION_IMPL *,
-		WT_RECONCILE *, const void *, size_t, bool *);
-static int  __rec_cell_build_ovfl(WT_SESSION_IMPL *, WT_RECONCILE *,
-		WT_REC_KV *, uint8_t,
-		wt_timestamp_t, wt_timestamp_t, uint64_t, uint64_t, uint64_t);
-static int  __rec_cell_build_val(WT_SESSION_IMPL *, WT_RECONCILE *,
-		const void *, size_t,
-		wt_timestamp_t, wt_timestamp_t, uint64_t, uint64_t, uint64_t);
 static void __rec_cleanup(WT_SESSION_IMPL *, WT_RECONCILE *);
-static int  __rec_col_fix(WT_SESSION_IMPL *, WT_RECONCILE *, WT_REF *);
-static int  __rec_col_fix_slvg(WT_SESSION_IMPL *,
-		WT_RECONCILE *, WT_REF *, WT_SALVAGE_COOKIE *);
-static int  __rec_col_int(WT_SESSION_IMPL *, WT_RECONCILE *, WT_REF *);
-static int  __rec_col_merge(WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
-static int  __rec_col_var(WT_SESSION_IMPL *,
-		WT_RECONCILE *, WT_REF *, WT_SALVAGE_COOKIE *);
 static int  __rec_destroy_session(WT_SESSION_IMPL *);
 static int  __rec_init(WT_SESSION_IMPL *,
 		WT_REF *, uint32_t, WT_SALVAGE_COOKIE *, void *);
 static int  __rec_las_wrapup(WT_SESSION_IMPL *, WT_RECONCILE *);
 static int  __rec_las_wrapup_err(WT_SESSION_IMPL *, WT_RECONCILE *);
 static int  __rec_root_write(WT_SESSION_IMPL *, WT_PAGE *, uint32_t);
-static int  __rec_row_int(WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
-static int  __rec_row_leaf(WT_SESSION_IMPL *,
-		WT_RECONCILE *, WT_REF *, WT_SALVAGE_COOKIE *);
-static int  __rec_row_leaf_insert(
-		WT_SESSION_IMPL *, WT_RECONCILE *, WT_INSERT *);
-static int  __rec_row_merge(WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
 static int  __rec_split_discard(WT_SESSION_IMPL *, WT_PAGE *);
 static int  __rec_split_row_promote(
 		WT_SESSION_IMPL *, WT_RECONCILE *, WT_ITEM *, uint8_t);
@@ -165,23 +140,23 @@ __wt_reconcile(WT_SESSION_IMPL *session, WT_REF *ref,
 	switch (page->type) {
 	case WT_PAGE_COL_FIX:
 		if (salvage != NULL)
-			ret = __rec_col_fix_slvg(session, r, ref, salvage);
+			ret = __wt_rec_col_fix_slvg(session, r, ref, salvage);
 		else
-			ret = __rec_col_fix(session, r, ref);
+			ret = __wt_rec_col_fix(session, r, ref);
 		break;
 	case WT_PAGE_COL_INT:
 		WT_WITH_PAGE_INDEX(session,
-		    ret = __rec_col_int(session, r, ref));
+		    ret = __wt_rec_col_int(session, r, ref));
 		break;
 	case WT_PAGE_COL_VAR:
-		ret = __rec_col_var(session, r, ref, salvage);
+		ret = __wt_rec_col_var(session, r, ref, salvage);
 		break;
 	case WT_PAGE_ROW_INT:
 		WT_WITH_PAGE_INDEX(session,
-		    ret = __rec_row_int(session, r, page));
+		    ret = __wt_rec_row_int(session, r, page));
 		break;
 	case WT_PAGE_ROW_LEAF:
-		ret = __rec_row_leaf(session, r, ref, salvage);
+		ret = __wt_rec_row_leaf(session, r, ref, salvage);
 		break;
 	default:
 		ret = __wt_illegal_value(session, page->type);
@@ -817,523 +792,6 @@ __rec_destroy_session(WT_SESSION_IMPL *session)
 }
 
 /*
- * WT_CHILD_RELEASE, WT_CHILD_RELEASE_ERR --
- *	Macros to clean up during internal-page reconciliation, releasing the
- *	hazard pointer we're holding on child pages.
- */
-#define	WT_CHILD_RELEASE(session, hazard, ref) do {			\
-	if (hazard) {							\
-		(hazard) = false;					\
-		WT_TRET(						\
-		    __wt_page_release(session, ref, WT_READ_NO_EVICT));	\
-	}								\
-} while (0)
-#define	WT_CHILD_RELEASE_ERR(session, hazard, ref) do {			\
-	WT_CHILD_RELEASE(session, hazard, ref);				\
-	WT_ERR(ret);							\
-} while (0)
-
-typedef enum {
-    WT_CHILD_IGNORE,				/* Ignored child */
-    WT_CHILD_MODIFIED,				/* Modified child */
-    WT_CHILD_ORIGINAL,				/* Original child */
-    WT_CHILD_PROXY				/* Deleted child: proxy */
-} WT_CHILD_STATE;
-
-/*
- * __rec_child_deleted --
- *	Handle pages with leaf pages in the WT_REF_DELETED state.
- */
-static int
-__rec_child_deleted(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *ref, WT_CHILD_STATE *statep)
-{
-	WT_PAGE_DELETED *page_del;
-
-	page_del = ref->page_del;
-
-	/*
-	 * Internal pages with child leaf pages in the WT_REF_DELETED state are
-	 * a special case during reconciliation.  First, if the deletion was a
-	 * result of a session truncate call, the deletion may not be visible to
-	 * us. In that case, we proceed as with any change not visible during
-	 * reconciliation by ignoring the change for the purposes of writing the
-	 * internal page.
-	 *
-	 * In this case, there must be an associated page-deleted structure, and
-	 * it holds the transaction ID we care about.
-	 *
-	 * In some cases, there had better not be any updates we can't see.
-	 *
-	 * A visible update to be in READY state (i.e. not in LOCKED or
-	 * PREPARED state), for truly visible to others.
-	 */
-	if (F_ISSET(r, WT_REC_VISIBILITY_ERR) && page_del != NULL &&
-	    __wt_page_del_active(session, ref, false))
-		WT_PANIC_RET(session, EINVAL,
-		    "reconciliation illegally skipped an update");
-
-	/*
-	 * Deal with any underlying disk blocks.
-	 *
-	 * First, check to see if there is an address associated with this leaf:
-	 * if there isn't, we're done, the underlying page is already gone.  If
-	 * the page still exists, check for any transactions in the system that
-	 * might want to see the page's state before it's deleted.
-	 *
-	 * If any such transactions exist, we cannot discard the underlying leaf
-	 * page to the block manager because the transaction may eventually read
-	 * it.  However, this write might be part of a checkpoint, and should we
-	 * recover to that checkpoint, we'll need to delete the leaf page, else
-	 * we'd leak it.  The solution is to write a proxy cell on the internal
-	 * page ensuring the leaf page is eventually discarded.
-	 *
-	 * If no such transactions exist, we can discard the leaf page to the
-	 * block manager and no cell needs to be written at all.  We do this
-	 * outside of the underlying tracking routines because this action is
-	 * permanent and irrevocable.  (Clearing the address means we've lost
-	 * track of the disk address in a permanent way.  This is safe because
-	 * there's no path to reading the leaf page again: if there's ever a
-	 * read into this part of the name space again, the cache read function
-	 * instantiates an entirely new page.)
-	 */
-	if (ref->addr != NULL && !__wt_page_del_active(session, ref, true)) {
-		/*
-		 * Minor memory cleanup: if a truncate call deleted this page
-		 * and we were ever forced to instantiate the page in memory,
-		 * we would have built a list of updates in the page reference
-		 * in order to be able to commit/rollback the truncate. We just
-		 * passed a visibility test, discard the update list.
-		 */
-		if (page_del != NULL) {
-			__wt_free(session, ref->page_del->update_list);
-			__wt_free(session, ref->page_del);
-		}
-
-		WT_RET(__wt_ref_block_free(session, ref));
-	}
-
-	/*
-	 * If the original page is gone, we can skip the slot on the internal
-	 * page.
-	 */
-	if (ref->addr == NULL) {
-		*statep = WT_CHILD_IGNORE;
-		return (0);
-	}
-
-	/*
-	 * Internal pages with deletes that aren't stable cannot be evicted, we
-	 * don't have sufficient information to restore the page's information
-	 * if subsequently read (we wouldn't know which transactions should see
-	 * the original page and which should see the deleted page).
-	 */
-	if (F_ISSET(r, WT_REC_EVICT))
-		return (__wt_set_return(session, EBUSY));
-
-	/*
-	 * If there are deleted child pages we can't discard immediately, keep
-	 * the page dirty so they are eventually freed.
-	 */
-	r->leave_dirty = true;
-
-	/*
-	 * If the original page cannot be freed, we need to keep a slot on the
-	 * page to reference it from the parent page.
-	 *
-	 * If the delete is not visible in this checkpoint, write the original
-	 * address normally. Otherwise, we have to write a proxy record.
-	 * If the delete state is not ready, then delete is not visible as it
-	 * is in prepared state.
-	 */
-	if (!__wt_page_del_active(session, ref, false))
-		*statep = WT_CHILD_PROXY;
-
-	return (0);
-}
-
-/*
- * __rec_child_modify --
- *	Return if the internal page's child references any modifications.
- */
-static int
-__rec_child_modify(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *ref, bool *hazardp, WT_CHILD_STATE *statep)
-{
-	WT_DECL_RET;
-	WT_PAGE_MODIFY *mod;
-
-	/* We may acquire a hazard pointer our caller must release. */
-	*hazardp = false;
-
-	/* Default to using the original child address. */
-	*statep = WT_CHILD_ORIGINAL;
-
-	/*
-	 * This function is called when walking an internal page to decide how
-	 * to handle child pages referenced by the internal page.
-	 *
-	 * Internal pages are reconciled for two reasons: first, when evicting
-	 * an internal page, second by the checkpoint code when writing internal
-	 * pages.  During eviction, all pages should be in the WT_REF_DISK or
-	 * WT_REF_DELETED state. During checkpoint, eviction that might affect
-	 * review of an internal page is prohibited, however, as the subtree is
-	 * not reserved for our exclusive use, there are other page states that
-	 * must be considered.
-	 */
-	for (;; __wt_yield()) {
-		switch (r->tested_ref_state = ref->state) {
-		case WT_REF_DISK:
-			/* On disk, not modified by definition. */
-			goto done;
-
-		case WT_REF_DELETED:
-			/*
-			 * The child is in a deleted state.
-			 *
-			 * It's possible the state could change underneath us as
-			 * the page is read in, and we can race between checking
-			 * for a deleted state and looking at the transaction ID
-			 * to see if the delete is visible to us.  Lock down the
-			 * structure.
-			 */
-			if (!__wt_atomic_casv32(
-			    &ref->state, WT_REF_DELETED, WT_REF_LOCKED))
-				break;
-			ret = __rec_child_deleted(session, r, ref, statep);
-			WT_REF_SET_STATE(ref, WT_REF_DELETED);
-			goto done;
-
-		case WT_REF_LOCKED:
-			/*
-			 * Locked.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 */
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			if (F_ISSET(r, WT_REC_EVICT))
-				return (__wt_set_return(session, EBUSY));
-
-			/*
-			 * If called during checkpoint, the child is being
-			 * considered by the eviction server or the child is a
-			 * truncated page being read.  The eviction may have
-			 * started before the checkpoint and so we must wait
-			 * for the eviction to be resolved.  I suspect we could
-			 * handle reads of truncated pages, but we can't
-			 * distinguish between the two and reads of truncated
-			 * pages aren't expected to be common.
-			 */
-			break;
-
-		case WT_REF_LIMBO:
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			/* FALLTHROUGH */
-		case WT_REF_LOOKASIDE:
-			/*
-			 * On disk or in cache with lookaside updates.
-			 *
-			 * We should never be here during eviction: active
-			 * child pages in an evicted page's subtree fails the
-			 * eviction attempt.
-			 */
-			if (F_ISSET(r, WT_REC_EVICT) &&
-			    __wt_page_las_active(session, ref)) {
-				WT_ASSERT(session, false);
-				return (__wt_set_return(session, EBUSY));
-			}
-
-			/*
-			 * A page evicted with lookaside entries may not have
-			 * an address, if no updates were visible to
-			 * reconciliation.  Any child pages in that state
-			 * should be ignored.
-			 */
-			if (ref->addr == NULL) {
-				*statep = WT_CHILD_IGNORE;
-				WT_CHILD_RELEASE(session, *hazardp, ref);
-			}
-			goto done;
-
-		case WT_REF_MEM:
-			/*
-			 * In memory.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 */
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			if (F_ISSET(r, WT_REC_EVICT))
-				return (__wt_set_return(session, EBUSY));
-
-			/*
-			 * If called during checkpoint, acquire a hazard pointer
-			 * so the child isn't evicted, it's an in-memory case.
-			 *
-			 * This call cannot return split/restart, we have a lock
-			 * on the parent which prevents a child page split.
-			 *
-			 * Set WT_READ_NO_WAIT because we're only interested in
-			 * the WT_REF's final state. Pages in transition might
-			 * change WT_REF state during our read, and then return
-			 * WT_NOTFOUND to us. In that case, loop and look again.
-			 */
-			ret = __wt_page_in(session, ref,
-			    WT_READ_CACHE | WT_READ_NO_EVICT |
-			    WT_READ_NO_GEN | WT_READ_NO_WAIT);
-			if (ret == WT_NOTFOUND) {
-				ret = 0;
-				break;
-			}
-			WT_RET(ret);
-			*hazardp = true;
-			goto in_memory;
-
-		case WT_REF_READING:
-			/*
-			 * Being read, not modified by definition.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 */
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			if (F_ISSET(r, WT_REC_EVICT))
-				return (__wt_set_return(session, EBUSY));
-			goto done;
-
-		case WT_REF_SPLIT:
-			/*
-			 * The page was split out from under us.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 *
-			 * We should never be here during checkpoint, dirty page
-			 * eviction is shutout during checkpoint, all splits in
-			 * process will have completed before we walk any pages
-			 * for checkpoint.
-			 */
-			WT_ASSERT(session, WT_REF_SPLIT != WT_REF_SPLIT);
-			return (__wt_set_return(session, EBUSY));
-
-		WT_ILLEGAL_VALUE(session, r->tested_ref_state);
-		}
-		WT_STAT_CONN_INCR(session, child_modify_blocked_page);
-	}
-
-in_memory:
-	/*
-	 * In-memory states: the child is potentially modified if the page's
-	 * modify structure has been instantiated. If the modify structure
-	 * exists and the page has actually been modified, set that state.
-	 * If that's not the case, we would normally use the original cell's
-	 * disk address as our reference, however there are two special cases,
-	 * both flagged by a missing block address.
-	 *
-	 * First, if forced to instantiate a deleted child page and it's never
-	 * modified, we end up here with a page that has a modify structure, no
-	 * modifications, and no disk address. Ignore those pages, they're not
-	 * modified and there is no reason to write the cell.
-	 *
-	 * Second, insert splits are permitted during checkpoint. When doing the
-	 * final checkpoint pass, we first walk the internal page's page-index
-	 * and write out any dirty pages we find, then we write out the internal
-	 * page in post-order traversal. If we found the split page in the first
-	 * step, it will have an address; if we didn't find the split page in
-	 * the first step, it won't have an address and we ignore it, it's not
-	 * part of the checkpoint.
-	 */
-	mod = ref->page->modify;
-	if (mod != NULL && mod->rec_result != 0)
-		*statep = WT_CHILD_MODIFIED;
-	else if (ref->addr == NULL) {
-		*statep = WT_CHILD_IGNORE;
-		WT_CHILD_RELEASE(session, *hazardp, ref);
-	}
-
-done:	WT_DIAGNOSTIC_YIELD;
-	return (ret);
-}
-
-/*
- * __rec_incr --
- *	Update the memory tracking structure for a set of new entries.
- */
-static inline void
-__rec_incr(WT_SESSION_IMPL *session, WT_RECONCILE *r, uint32_t v, size_t size)
-{
-	/*
-	 * The buffer code is fragile and prone to off-by-one errors -- check
-	 * for overflow in diagnostic mode.
-	 */
-	WT_ASSERT(session, r->space_avail >= size);
-	WT_ASSERT(session, WT_BLOCK_FITS(r->first_free, size,
-	    r->cur_ptr->image.mem, r->cur_ptr->image.memsize));
-
-	r->entries += v;
-	r->space_avail -= size;
-	r->first_free += size;
-
-	/*
-	 * If offset for the minimum split size boundary is not set, we have not
-	 * yet reached the minimum boundary, reduce the space available for it.
-	 */
-	if (r->cur_ptr->min_offset == 0) {
-		if (r->min_space_avail >= size)
-			r->min_space_avail -= size;
-		else
-			r->min_space_avail = 0;
-	}
-}
-
-/*
- * __rec_image_copy --
- *	Copy a key/value cell and buffer pair into the new image.
- */
-static inline void
-__rec_image_copy(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_REC_KV *kv)
-{
-	size_t len;
-	uint8_t *p, *t;
-
-	/*
-	 * If there's only one chunk of data to copy (because the cell and data
-	 * are being copied from the original disk page), the cell length won't
-	 * be set, the WT_ITEM data/length will reference the data to be copied.
-	 *
-	 * WT_CELLs are typically small, 1 or 2 bytes -- don't call memcpy, do
-	 * the copy in-line.
-	 */
-	for (p = r->first_free,
-	    t = (uint8_t *)&kv->cell, len = kv->cell_len; len > 0; --len)
-		*p++ = *t++;
-
-	/* The data can be quite large -- call memcpy. */
-	if (kv->buf.size != 0)
-		memcpy(p, kv->buf.data, kv->buf.size);
-
-	WT_ASSERT(session, kv->len == kv->cell_len + kv->buf.size);
-	__rec_incr(session, r, 1, kv->len);
-}
-
-/*
- * __rec_dict_replace --
- *	Check for a dictionary match.
- */
-static int
-__rec_dict_replace(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    wt_timestamp_t start_ts, wt_timestamp_t stop_ts,
-    uint64_t start_txn, uint64_t stop_txn, uint64_t rle, WT_REC_KV *val)
-{
-	WT_REC_DICTIONARY *dp;
-	uint64_t offset;
-
-	/*
-	 * We optionally create a dictionary of values and only write a unique
-	 * value once per page, using a special "copy" cell for all subsequent
-	 * copies of the value.  We have to do the cell build and resolution at
-	 * this low level because we need physical cell offsets for the page.
-	 *
-	 * Sanity check: short-data cells can be smaller than dictionary-copy
-	 * cells.  If the data is already small, don't bother doing the work.
-	 * This isn't just work avoidance: on-page cells can't grow as a result
-	 * of writing a dictionary-copy cell, the reconciliation functions do a
-	 * split-boundary test based on the size required by the value's cell;
-	 * if we grow the cell after that test we'll potentially write off the
-	 * end of the buffer's memory.
-	 */
-	if (val->buf.size <= WT_INTPACK32_MAXSIZE)
-		return (0);
-	WT_RET(__wt_rec_dictionary_lookup(session, r, val, &dp));
-	if (dp == NULL)
-		return (0);
-
-	/*
-	 * If the dictionary offset isn't set, we're creating a new entry in the
-	 * dictionary, set its location.
-	 *
-	 * If the dictionary offset is set, we have a matching value. Create a
-	 * copy cell instead.
-	 */
-	if (dp->offset == 0)
-		dp->offset = WT_PTRDIFF32(r->first_free, r->cur_ptr->image.mem);
-	else {
-		/*
-		 * The offset is the byte offset from this cell to the previous,
-		 * matching cell, NOT the byte offset from the beginning of the
-		 * page.
-		 */
-		offset = (uint64_t)WT_PTRDIFF(r->first_free,
-		    (uint8_t *)r->cur_ptr->image.mem + dp->offset);
-		val->len = val->cell_len = __wt_cell_pack_copy(session,
-		    &val->cell,
-		    start_ts, stop_ts, start_txn, stop_txn, rle, offset);
-		val->buf.data = NULL;
-		val->buf.size = 0;
-	}
-	return (0);
-}
-
-/*
- * __rec_key_state_update --
- *	Update prefix and suffix compression based on the last key.
- */
-static inline void
-__rec_key_state_update(WT_RECONCILE *r, bool ovfl_key)
-{
-	WT_ITEM *a;
-
-	/*
-	 * If writing an overflow key onto the page, don't update the "last key"
-	 * value, and leave the state of prefix compression alone.  (If we are
-	 * currently doing prefix compression, we have a key state which will
-	 * continue to work, we're just skipping the key just created because
-	 * it's an overflow key and doesn't participate in prefix compression.
-	 * If we are not currently doing prefix compression, we can't start, an
-	 * overflow key doesn't give us any state.)
-	 *
-	 * Additionally, if we wrote an overflow key onto the page, turn off the
-	 * suffix compression of row-store internal node keys.  (When we split,
-	 * "last key" is the largest key on the previous page, and "cur key" is
-	 * the first key on the next page, which is being promoted.  In some
-	 * cases we can discard bytes from the "cur key" that are not needed to
-	 * distinguish between the "last key" and "cur key", compressing the
-	 * size of keys on internal nodes.  If we just built an overflow key,
-	 * we're not going to update the "last key", making suffix compression
-	 * impossible for the next key. Alternatively, we could remember where
-	 * the last key was on the page, detect it's an overflow key, read it
-	 * from disk and do suffix compression, but that's too much work for an
-	 * unlikely event.)
-	 *
-	 * If we're not writing an overflow key on the page, update the last-key
-	 * value and turn on both prefix and suffix compression.
-	 */
-	if (ovfl_key)
-		r->key_sfx_compress = false;
-	else {
-		a = r->cur;
-		r->cur = r->last;
-		r->last = a;
-
-		r->key_pfx_compress = r->key_pfx_compress_conf;
-		r->key_sfx_compress = r->key_sfx_compress_conf;
-	}
-}
-
-/*
- * Macros from fixed-length entries to/from bytes.
- */
-#define	WT_FIX_BYTES_TO_ENTRIES(btree, bytes)				\
-    ((uint32_t)((((bytes) * 8) / (btree)->bitcnt)))
-#define	WT_FIX_ENTRIES_TO_BYTES(btree, entries)				\
-	((uint32_t)WT_ALIGN((entries) * (btree)->bitcnt, 8))
-
-/*
  * __rec_leaf_page_max --
  *	Figure out the maximum leaf page size for the reconciliation.
  */
@@ -1399,42 +857,6 @@ __rec_leaf_page_max(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 	return (page_size * 2);
 }
 
-#define	WT_CROSSING_MIN_BND(r, next_len)				\
-	((r)->cur_ptr->min_offset == 0 &&				\
-	    (next_len) > (r)->min_space_avail)
-#define	WT_CROSSING_SPLIT_BND(r, next_len) ((next_len) > (r)->space_avail)
-#define	WT_CHECK_CROSSING_BND(r, next_len)				\
-	(WT_CROSSING_MIN_BND(r, next_len) || WT_CROSSING_SPLIT_BND(r, next_len))
-
-/*
- * __rec_need_split --
- *	Check whether adding some bytes to the page requires a split.
- */
-static bool
-__rec_need_split(WT_RECONCILE *r, size_t len)
-{
-	/*
-	 * In the case of a row-store leaf page, trigger a split if a threshold
-	 * number of saved updates is reached. This allows pages to split for
-	 * update/restore and lookaside eviction when there is no visible data
-	 * causing the disk image to grow.
-	 *
-	 * In the case of small pages or large keys, we might try to split when
-	 * a page has no updates or entries, which isn't possible. To consider
-	 * update/restore or lookaside information, require either page entries
-	 * or updates that will be attached to the image. The limit is one of
-	 * either, but it doesn't make sense to create pages or images with few
-	 * entries or updates, even where page sizes are small (especially as
-	 * updates that will eventually become overflow items can throw off our
-	 * calculations). Bound the combination at something reasonable.
-	 */
-	if (r->page->type == WT_PAGE_ROW_LEAF && r->entries + r->supd_next > 10)
-		len += r->supd_memsize;
-
-	/* Check for the disk image crossing a boundary. */
-	return (WT_CHECK_CROSSING_BND(r, len));
-}
-
 /*
  * __wt_split_page_size --
  *	Given a split percentage, calculate split page size in bytes.
@@ -1466,56 +888,6 @@ __wt_split_page_size(int split_pct, uint32_t maxpagesize, uint32_t allocsize)
 		split_size = (uint32_t)((a * (u_int)split_pct) / 100);
 
 	return (split_size);
-}
-
-/*
- * __rec_addr_ts_init --
- *	Initialize an address timestamp triplet.
- */
-static void
-__rec_addr_ts_init(WT_RECONCILE *r, wt_timestamp_t *oldest_start_tsp,
-    wt_timestamp_t *newest_durable_ts, wt_timestamp_t *newest_stop_tsp,
-    uint64_t *oldest_start_txnp, uint64_t *newest_stop_txnp)
-{
-	/*
-	 * If the page format supports address timestamps (and not fixed-length
-	 * column-store, where we don't maintain timestamps at all), set the
-	 * oldest/newest timestamps to values at the end of their expected range
-	 * so they're corrected as we process key/value items. Otherwise, set
-	 * the oldest/newest timestamps to simple durability.
-	 */
-	*oldest_start_tsp = WT_TS_MAX;
-	*newest_durable_ts = *newest_stop_tsp = WT_TS_NONE;
-	*oldest_start_txnp = WT_TXN_MAX;
-	*newest_stop_txnp = WT_TXN_NONE;
-	if (!__wt_process.page_version_ts || r->page->type == WT_PAGE_COL_FIX) {
-		*oldest_start_tsp = *newest_durable_ts = WT_TS_NONE;
-		*newest_stop_tsp = WT_TS_MAX;
-		*oldest_start_txnp = WT_TXN_NONE;
-		*newest_stop_txnp = WT_TXN_MAX;
-	}
-}
-
-/*
- * __rec_addr_ts_update --
- *	Update the chunk's timestamp information.
- */
-static inline void
-__rec_addr_ts_update(WT_RECONCILE *r, wt_timestamp_t oldest_start_ts,
-    wt_timestamp_t newest_durable_ts, wt_timestamp_t newest_stop_ts,
-    uint64_t oldest_start_txn, uint64_t newest_stop_txn)
-{
-	r->cur_ptr->oldest_start_ts =
-	    WT_MIN(oldest_start_ts, r->cur_ptr->oldest_start_ts);
-	r->cur_ptr->newest_durable_ts =
-	    WT_MAX(newest_durable_ts, r->cur_ptr->newest_durable_ts);
-	r->cur_ptr->newest_stop_ts =
-	    WT_MAX(newest_stop_ts, r->cur_ptr->newest_stop_ts);
-
-	r->cur_ptr->oldest_start_txn =
-	    WT_MIN(oldest_start_txn, r->cur_ptr->oldest_start_txn);
-	r->cur_ptr->newest_stop_txn =
-	    WT_MAX(newest_stop_txn, r->cur_ptr->newest_stop_txn);
 }
 
 /*
@@ -1563,11 +935,11 @@ __rec_split_chunk_init(WT_SESSION_IMPL *session,
 }
 
 /*
- * __rec_split_init --
+ * __wt_rec_split_init --
  *	Initialization for the reconciliation split functions.
  */
-static int
-__rec_split_init(WT_SESSION_IMPL *session,
+int
+__wt_rec_split_init(WT_SESSION_IMPL *session,
     WT_RECONCILE *r, WT_PAGE *page, uint64_t recno, uint64_t max)
 {
 	WT_BM *bm;
@@ -1868,12 +1240,12 @@ __rec_split_grow(WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t add_len)
 }
 
 /*
- * __rec_split --
+ * __wt_rec_split --
  *	Handle the page reconciliation bookkeeping.  (Did you know "bookkeeper"
  *	has 3 doubled letters in a row?  Sweet-tooth does, too.)
  */
-static int
-__rec_split(WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t next_len)
+int
+__wt_rec_split(WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t next_len)
 {
 	WT_BTREE *btree;
 	WT_REC_CHUNK *tmp;
@@ -1970,12 +1342,12 @@ done:  	/*
 }
 
 /*
- * __rec_split_crossing_bnd --
+ * __wt_rec_split_crossing_bnd --
  * 	Save the details for the minimum split size boundary or call for a
  * 	split.
  */
-static inline int
-__rec_split_crossing_bnd(
+int
+__wt_rec_split_crossing_bnd(
     WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t next_len)
 {
 	WT_ASSERT(session, __rec_need_split(r, next_len));
@@ -2021,7 +1393,7 @@ __rec_split_crossing_bnd(
 	}
 
 	/* We are crossing a split boundary */
-	return (__rec_split(session, r, next_len));
+	return (__wt_rec_split(session, r, next_len));
 }
 
 /*
@@ -2151,11 +1523,11 @@ __rec_split_finish_process_prev(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 }
 
 /*
- * __rec_split_finish --
+ * __wt_rec_split_finish --
  *	Finish processing a page.
  */
-static int
-__rec_split_finish(WT_SESSION_IMPL *session, WT_RECONCILE *r)
+int
+__wt_rec_split_finish(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 {
 	/*
 	 * We're done reconciling, write the final page. We may arrive here with
@@ -2754,7 +2126,7 @@ __wt_bulk_init(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
 
 	recno = btree->type == BTREE_ROW ? WT_RECNO_OOB : 1;
 
-	return (__rec_split_init(session,
+	return (__wt_rec_split_init(session,
 	    r, cbulk->leaf, recno, btree->maxleafpage_precomp));
 }
 
@@ -2788,7 +2160,7 @@ __wt_bulk_wrapup(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
 		break;
 	}
 
-	WT_RET(__rec_split_finish(session, r));
+	WT_RET(__wt_rec_split_finish(session, r));
 	WT_RET(__rec_write_wrapup(session, r, r->page));
 	__rec_write_page_status(session, r);
 
@@ -2799,2166 +2171,6 @@ __wt_bulk_wrapup(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
 
 	__rec_cleanup(session, r);
 	__rec_destroy(session, &cbulk->reconcile);
-
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_row --
- *	Row-store bulk insert.
- */
-int
-__wt_bulk_insert_row(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
-{
-	WT_BTREE *btree;
-	WT_CURSOR *cursor;
-	WT_RECONCILE *r;
-	WT_REC_KV *key, *val;
-	bool ovfl_key;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-	cursor = &cbulk->cbt.iface;
-
-	key = &r->k;
-	val = &r->v;
-	WT_RET(__rec_cell_build_leaf_key(session, r,	/* Build key cell */
-	    cursor->key.data, cursor->key.size, &ovfl_key));
-	WT_RET(__rec_cell_build_val(session, r,		/* Build value cell */
-	    cursor->value.data, cursor->value.size,
-	    WT_TS_NONE, WT_TS_MAX, WT_TXN_NONE, WT_TXN_MAX, 0));
-
-	/* Boundary: split or write the page. */
-	if (WT_CROSSING_SPLIT_BND(r, key->len + val->len)) {
-		/*
-		 * Turn off prefix compression until a full key written to the
-		 * new page, and (unless already working with an overflow key),
-		 * rebuild the key without compression.
-		 */
-		if (r->key_pfx_compress_conf) {
-			r->key_pfx_compress = false;
-			if (!ovfl_key)
-				WT_RET(__rec_cell_build_leaf_key(
-				    session, r, NULL, 0, &ovfl_key));
-		}
-		WT_RET(__rec_split_crossing_bnd(
-		    session, r, key->len + val->len));
-	}
-
-	/* Copy the key/value pair onto the page. */
-	__rec_image_copy(session, r, key);
-	if (val->len == 0)
-		r->any_empty_value = true;
-	else {
-		r->all_empty_value = false;
-		if (btree->dictionary)
-			WT_RET(__rec_dict_replace(session, r,
-			    WT_TS_NONE, WT_TS_MAX, WT_TXN_NONE, WT_TXN_MAX,
-			    0, val));
-		__rec_image_copy(session, r, val);
-	}
-	__rec_addr_ts_update(r,
-	    WT_TS_NONE, WT_TS_NONE, WT_TS_MAX, WT_TXN_NONE, WT_TXN_MAX);
-
-	/* Update compression state. */
-	__rec_key_state_update(r, ovfl_key);
-
-	return (0);
-}
-
-/*
- * __rec_col_fix_bulk_insert_split_check --
- *	Check if a bulk-loaded fixed-length column store page needs to split.
- */
-static inline int
-__rec_col_fix_bulk_insert_split_check(WT_CURSOR_BULK *cbulk)
-{
-	WT_BTREE *btree;
-	WT_RECONCILE *r;
-	WT_SESSION_IMPL *session;
-
-	session = (WT_SESSION_IMPL *)cbulk->cbt.iface.session;
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-
-	if (cbulk->entry == cbulk->nrecs) {
-		if (cbulk->entry != 0) {
-			/*
-			 * If everything didn't fit, update the counters and
-			 * split.
-			 *
-			 * Boundary: split or write the page.
-			 *
-			 * No need to have a minimum split size boundary, all
-			 * pages are filled 100% except the last, allowing it to
-			 * grow in the future.
-			 */
-			__rec_incr(session, r, cbulk->entry,
-			    __bitstr_size(
-			    (size_t)cbulk->entry * btree->bitcnt));
-			WT_RET(__rec_split(session, r, 0));
-		}
-		cbulk->entry = 0;
-		cbulk->nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail);
-	}
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_fix --
- *	Fixed-length column-store bulk insert.
- */
-int
-__wt_bulk_insert_fix(
-    WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk, bool deleted)
-{
-	WT_BTREE *btree;
-	WT_CURSOR *cursor;
-	WT_RECONCILE *r;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-	cursor = &cbulk->cbt.iface;
-
-	WT_RET(__rec_col_fix_bulk_insert_split_check(cbulk));
-	__bit_setv(r->first_free, cbulk->entry,
-	    btree->bitcnt, deleted ? 0 : ((uint8_t *)cursor->value.data)[0]);
-	++cbulk->entry;
-	++r->recno;
-
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_fix_bitmap --
- *	Fixed-length column-store bulk insert.
- */
-int
-__wt_bulk_insert_fix_bitmap(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
-{
-	WT_BTREE *btree;
-	WT_CURSOR *cursor;
-	WT_RECONCILE *r;
-	uint32_t entries, offset, page_entries, page_size;
-	const uint8_t *data;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-	cursor = &cbulk->cbt.iface;
-
-	if (((r->recno - 1) * btree->bitcnt) & 0x7)
-		WT_RET_MSG(session, EINVAL,
-		    "Bulk bitmap load not aligned on a byte boundary");
-	for (data = cursor->value.data,
-	    entries = (uint32_t)cursor->value.size;
-	    entries > 0;
-	    entries -= page_entries, data += page_size) {
-		WT_RET(__rec_col_fix_bulk_insert_split_check(cbulk));
-
-		page_entries = WT_MIN(entries, cbulk->nrecs - cbulk->entry);
-		page_size = __bitstr_size(page_entries * btree->bitcnt);
-		offset = __bitstr_size(cbulk->entry * btree->bitcnt);
-		memcpy(r->first_free + offset, data, page_size);
-		cbulk->entry += page_entries;
-		r->recno += page_entries;
-	}
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_var --
- *	Variable-length column-store bulk insert.
- */
-int
-__wt_bulk_insert_var(
-    WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk, bool deleted)
-{
-	WT_BTREE *btree;
-	WT_RECONCILE *r;
-	WT_REC_KV *val;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-
-	val = &r->v;
-	if (deleted) {
-		val->cell_len = __wt_cell_pack_del(session, &val->cell,
-		    WT_TS_NONE, WT_TS_MAX, WT_TXN_NONE, WT_TXN_MAX, cbulk->rle);
-		val->buf.data = NULL;
-		val->buf.size = 0;
-		val->len = val->cell_len;
-	} else
-		/*
-		 * Store the bulk cursor's last buffer, not the current value,
-		 * we're tracking duplicates, which means we want the previous
-		 * value seen, not the current value.
-		 */
-		WT_RET(__rec_cell_build_val(session, r,
-		    cbulk->last.data, cbulk->last.size,
-		    WT_TS_NONE, WT_TS_MAX, WT_TXN_NONE, WT_TXN_MAX,
-		    cbulk->rle));
-
-	/* Boundary: split or write the page. */
-	if (WT_CROSSING_SPLIT_BND(r, val->len))
-		WT_RET(__rec_split_crossing_bnd(session, r, val->len));
-
-	/* Copy the value onto the page. */
-	if (btree->dictionary)
-		WT_RET(__rec_dict_replace(session, r,
-		    WT_TS_NONE, WT_TS_MAX, WT_TXN_NONE, WT_TXN_MAX,
-		    cbulk->rle, val));
-	__rec_image_copy(session, r, val);
-	__rec_addr_ts_update(r,
-	    WT_TS_NONE, WT_TS_NONE, WT_TS_MAX, WT_TXN_NONE, WT_TXN_MAX);
-
-	/* Update the starting record number in case we split. */
-	r->recno += cbulk->rle;
-
-	return (0);
-}
-
-/*
- * __rec_col_int --
- *	Reconcile a column-store internal page.
- */
-static int
-__rec_col_int(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_REF *pageref)
-{
-	WT_ADDR *addr;
-	WT_BTREE *btree;
-	WT_CELL_UNPACK *vpack, _vpack;
-	WT_CHILD_STATE state;
-	WT_DECL_RET;
-	WT_PAGE *child, *page;
-	WT_REC_KV *val;
-	WT_REF *ref;
-	wt_timestamp_t oldest_start_ts, newest_durable_ts, newest_stop_ts;
-	uint64_t oldest_start_txn, newest_stop_txn;
-	bool hazard;
-
-	btree = S2BT(session);
-	page = pageref->page;
-	child = NULL;
-	hazard = false;
-
-	val = &r->v;
-	vpack = &_vpack;
-
-	WT_RET(__rec_split_init(session,
-	    r, page, pageref->ref_recno, btree->maxintlpage_precomp));
-
-	/* For each entry in the in-memory page... */
-	WT_INTL_FOREACH_BEGIN(session, page, ref) {
-		/* Update the starting record number in case we split. */
-		r->recno = ref->ref_recno;
-
-		/*
-		 * Modified child.
-		 * The page may be emptied or internally created during a split.
-		 * Deleted/split pages are merged into the parent and discarded.
-		 */
-		WT_ERR(__rec_child_modify(session, r, ref, &hazard, &state));
-		addr = NULL;
-		child = ref->page;
-
-		switch (state) {
-		case WT_CHILD_IGNORE:
-			/* Ignored child. */
-			WT_CHILD_RELEASE_ERR(session, hazard, ref);
-			continue;
-
-		case WT_CHILD_MODIFIED:
-			/*
-			 * Modified child. Empty pages are merged into the
-			 * parent and discarded.
-			 */
-			switch (child->modify->rec_result) {
-			case WT_PM_REC_EMPTY:
-				/*
-				 * Column-store pages are almost never empty, as
-				 * discarding a page would remove a chunk of the
-				 * name space.  The exceptions are pages created
-				 * when the tree is created, and never filled.
-				 */
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_MULTIBLOCK:
-				WT_ERR(__rec_col_merge(session, r, child));
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_REPLACE:
-				addr = &child->modify->mod_replace;
-				break;
-			WT_ILLEGAL_VALUE_ERR(
-			    session, child->modify->rec_result);
-			}
-			break;
-		case WT_CHILD_ORIGINAL:
-			/* Original child. */
-			break;
-		case WT_CHILD_PROXY:
-			/*
-			 * Deleted child where we write a proxy cell, not yet
-			 * supported for column-store.
-			 */
-			WT_ERR(__wt_illegal_value(session, state));
-		}
-
-		/*
-		 * Build the value cell.  The child page address is in one of 3
-		 * places: if the page was replaced, the page's modify structure
-		 * references it and we built the value cell just above in the
-		 * switch statement.  Else, the WT_REF->addr reference points to
-		 * an on-page cell or an off-page WT_ADDR structure: if it's an
-		 * on-page cell and we copy it from the page, else build a new
-		 * cell.
-		 */
-		if (addr == NULL && __wt_off_page(page, ref->addr))
-			addr = ref->addr;
-		if (addr == NULL) {
-			__wt_cell_unpack(session, page, ref->addr, vpack);
-			val->buf.data = ref->addr;
-			val->buf.size = __wt_cell_total_len(vpack);
-			val->cell_len = 0;
-			val->len = val->buf.size;
-			oldest_start_ts = vpack->oldest_start_ts;
-			newest_durable_ts = vpack->newest_durable_ts;
-			newest_stop_ts = vpack->newest_stop_ts;
-			oldest_start_txn = vpack->oldest_start_txn;
-			newest_stop_txn = vpack->newest_stop_txn;
-		} else {
-			__rec_cell_build_addr(
-			    session, r, addr, false, ref->ref_recno);
-			oldest_start_ts = addr->oldest_start_ts;
-			newest_durable_ts = addr->newest_durable_ts;
-			newest_stop_ts = addr->newest_stop_ts;
-			oldest_start_txn = addr->oldest_start_txn;
-			newest_stop_txn = addr->newest_stop_txn;
-		}
-		WT_CHILD_RELEASE_ERR(session, hazard, ref);
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, val->len))
-			WT_ERR(__rec_split_crossing_bnd(session, r, val->len));
-
-		/* Copy the value onto the page. */
-		__rec_image_copy(session, r, val);
-		__rec_addr_ts_update(r,
-		    oldest_start_ts, newest_durable_ts, newest_stop_ts,
-		    oldest_start_txn, newest_stop_txn);
-	} WT_INTL_FOREACH_END;
-
-	/* Write the remnant page. */
-	return (__rec_split_finish(session, r));
-
-err:	WT_CHILD_RELEASE(session, hazard, ref);
-	return (ret);
-}
-
-/*
- * __rec_col_merge --
- *	Merge in a split page.
- */
-static int
-__rec_col_merge(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_PAGE *page)
-{
-	WT_ADDR *addr;
-	WT_MULTI *multi;
-	WT_PAGE_MODIFY *mod;
-	WT_REC_KV *val;
-	uint32_t i;
-
-	mod = page->modify;
-
-	val = &r->v;
-
-	/* For each entry in the split array... */
-	for (multi = mod->mod_multi,
-	    i = 0; i < mod->mod_multi_entries; ++multi, ++i) {
-		/* Update the starting record number in case we split. */
-		r->recno = multi->key.recno;
-
-		/* Build the value cell. */
-		addr = &multi->addr;
-		__rec_cell_build_addr(session, r, addr, false, r->recno);
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, val->len))
-			WT_RET(__rec_split_crossing_bnd(session, r, val->len));
-
-		/* Copy the value onto the page. */
-		__rec_image_copy(session, r, val);
-		__rec_addr_ts_update(r, addr->oldest_start_ts,
-		    addr->newest_durable_ts, addr->newest_stop_ts,
-		    addr->oldest_start_txn, addr->newest_stop_txn);
-	}
-	return (0);
-}
-
-/*
- * __rec_col_fix --
- *	Reconcile a fixed-width, column-store leaf page.
- */
-static int
-__rec_col_fix(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_REF *pageref)
-{
-	WT_BTREE *btree;
-	WT_INSERT *ins;
-	WT_PAGE *page;
-	WT_UPDATE *upd;
-	WT_UPDATE_SELECT upd_select;
-	uint64_t recno;
-	uint32_t entry, nrecs;
-
-	btree = S2BT(session);
-	page = pageref->page;
-
-	WT_RET(__rec_split_init(
-	    session, r, page, pageref->ref_recno, btree->maxleafpage));
-
-	/* Copy the original, disk-image bytes into place. */
-	memcpy(r->first_free, page->pg_fix_bitf,
-	    __bitstr_size((size_t)page->entries * btree->bitcnt));
-
-	/* Update any changes to the original on-page data items. */
-	WT_SKIP_FOREACH(ins, WT_COL_UPDATE_SINGLE(page)) {
-		WT_RET(__wt_rec_upd_select(
-		    session, r, ins, NULL, NULL, &upd_select));
-		upd = upd_select.upd;
-		if (upd != NULL)
-			__bit_setv(r->first_free,
-			    WT_INSERT_RECNO(ins) - pageref->ref_recno,
-			    btree->bitcnt, *upd->data);
-	}
-
-	/* Calculate the number of entries per page remainder. */
-	entry = page->entries;
-	nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail) - page->entries;
-	r->recno += entry;
-
-	/* Walk any append list. */
-	for (ins =
-	    WT_SKIP_FIRST(WT_COL_APPEND(page));; ins = WT_SKIP_NEXT(ins)) {
-		if (ins == NULL) {
-			/*
-			 * If the page split, instantiate any missing records in
-			 * the page's name space. (Imagine record 98 is
-			 * transactionally visible, 99 wasn't created or is not
-			 * yet visible, 100 is visible. Then the page splits and
-			 * record 100 moves to another page. When we reconcile
-			 * the original page, we write record 98, then we don't
-			 * see record 99 for whatever reason. If we've moved
-			 * record 100, we don't know to write a deleted record
-			 * 99 on the page.)
-			 *
-			 * The record number recorded during the split is the
-			 * first key on the split page, that is, one larger than
-			 * the last key on this page, we have to decrement it.
-			 */
-			if ((recno =
-			    page->modify->mod_col_split_recno) == WT_RECNO_OOB)
-				break;
-			recno -= 1;
-
-			/*
-			 * The following loop assumes records to write, and the
-			 * previous key might have been visible.
-			 */
-			if (r->recno > recno)
-				break;
-			upd = NULL;
-		} else {
-			WT_RET(__wt_rec_upd_select(
-			    session, r, ins, NULL, NULL, &upd_select));
-			upd = upd_select.upd;
-			recno = WT_INSERT_RECNO(ins);
-		}
-		for (;;) {
-			/*
-			 * The application may have inserted records which left
-			 * gaps in the name space.
-			 */
-			for (;
-			    nrecs > 0 && r->recno < recno;
-			    --nrecs, ++entry, ++r->recno)
-				__bit_setv(
-				    r->first_free, entry, btree->bitcnt, 0);
-
-			if (nrecs > 0) {
-				__bit_setv(r->first_free, entry, btree->bitcnt,
-				    upd == NULL ? 0 : *upd->data);
-				--nrecs;
-				++entry;
-				++r->recno;
-				break;
-			}
-
-			/*
-			 * If everything didn't fit, update the counters and
-			 * split.
-			 *
-			 * Boundary: split or write the page.
-			 *
-			 * No need to have a minimum split size boundary, all
-			 * pages are filled 100% except the last, allowing it to
-			 * grow in the future.
-			 */
-			__rec_incr(session, r, entry,
-			    __bitstr_size((size_t)entry * btree->bitcnt));
-			WT_RET(__rec_split(session, r, 0));
-
-			/* Calculate the number of entries per page. */
-			entry = 0;
-			nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail);
-		}
-
-		/*
-		 * Execute this loop once without an insert item to catch any
-		 * missing records due to a split, then quit.
-		 */
-		if (ins == NULL)
-			break;
-	}
-
-	/* Update the counters. */
-	__rec_incr(
-	    session, r, entry, __bitstr_size((size_t)entry * btree->bitcnt));
-
-	/* Write the remnant page. */
-	return (__rec_split_finish(session, r));
-}
-
-/*
- * __rec_col_fix_slvg --
- *	Reconcile a fixed-width, column-store leaf page created during salvage.
- */
-static int
-__rec_col_fix_slvg(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *pageref, WT_SALVAGE_COOKIE *salvage)
-{
-	WT_BTREE *btree;
-	WT_PAGE *page;
-	uint64_t page_start, page_take;
-	uint32_t entry, nrecs;
-
-	btree = S2BT(session);
-	page = pageref->page;
-
-	/*
-	 * !!!
-	 * It's vanishingly unlikely and probably impossible for fixed-length
-	 * column-store files to have overlapping key ranges.  It's possible
-	 * for an entire key range to go missing (if a page is corrupted and
-	 * lost), but because pages can't split, it shouldn't be possible to
-	 * find pages where the key ranges overlap.  That said, we check for
-	 * it during salvage and clean up after it here because it doesn't
-	 * cost much and future column-store formats or operations might allow
-	 * for fixed-length format ranges to overlap during salvage, and I
-	 * don't want to have to retrofit the code later.
-	 */
-	WT_RET(__rec_split_init(
-	    session, r, page, pageref->ref_recno, btree->maxleafpage));
-
-	/* We may not be taking all of the entries on the original page. */
-	page_take = salvage->take == 0 ? page->entries : salvage->take;
-	page_start = salvage->skip == 0 ? 0 : salvage->skip;
-
-	/* Calculate the number of entries per page. */
-	entry = 0;
-	nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail);
-
-	for (; nrecs > 0 && salvage->missing > 0;
-	    --nrecs, --salvage->missing, ++entry)
-		__bit_setv(r->first_free, entry, btree->bitcnt, 0);
-
-	for (; nrecs > 0 && page_take > 0;
-	    --nrecs, --page_take, ++page_start, ++entry)
-		__bit_setv(r->first_free, entry, btree->bitcnt,
-		    __bit_getv(page->pg_fix_bitf,
-			(uint32_t)page_start, btree->bitcnt));
-
-	r->recno += entry;
-	__rec_incr(session, r, entry,
-	    __bitstr_size((size_t)entry * btree->bitcnt));
-
-	/*
-	 * We can't split during salvage -- if everything didn't fit, it's
-	 * all gone wrong.
-	 */
-	if (salvage->missing != 0 || page_take != 0)
-		WT_PANIC_RET(session, WT_PANIC,
-		    "%s page too large, attempted split during salvage",
-		    __wt_page_type_string(page->type));
-
-	/* Write the page. */
-	return (__rec_split_finish(session, r));
-}
-
-/*
- * __rec_col_var_helper --
- *	Create a column-store variable length record cell and write it onto a
- *	page.
- */
-static int
-__rec_col_var_helper(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    WT_SALVAGE_COOKIE *salvage, WT_ITEM *value,
-    wt_timestamp_t start_ts, wt_timestamp_t durable_ts, wt_timestamp_t stop_ts,
-    uint64_t start_txn, uint64_t stop_txn,
-    uint64_t rle, bool deleted, bool overflow_type)
-{
-	WT_BTREE *btree;
-	WT_REC_KV *val;
-
-	btree = S2BT(session);
-	val = &r->v;
-
-	/*
-	 * Occasionally, salvage needs to discard records from the beginning or
-	 * end of the page, and because the items may be part of a RLE cell, do
-	 * the adjustments here. It's not a mistake we don't bother telling
-	 * our caller we've handled all the records from the page we care about,
-	 * and can quit processing the page: salvage is a rare operation and I
-	 * don't want to complicate our caller's loop.
-	 */
-	if (salvage != NULL) {
-		if (salvage->done)
-			return (0);
-		if (salvage->skip != 0) {
-			if (rle <= salvage->skip) {
-				salvage->skip -= rle;
-				return (0);
-			}
-			rle -= salvage->skip;
-			salvage->skip = 0;
-		}
-		if (salvage->take != 0) {
-			if (rle <= salvage->take)
-				salvage->take -= rle;
-			else {
-				rle = salvage->take;
-				salvage->take = 0;
-			}
-			if (salvage->take == 0)
-				salvage->done = true;
-		}
-	}
-
-	if (deleted) {
-		val->cell_len = __wt_cell_pack_del(session,
-		    &val->cell, start_ts, stop_ts, start_txn, stop_txn, rle);
-		val->buf.data = NULL;
-		val->buf.size = 0;
-		val->len = val->cell_len;
-	} else if (overflow_type) {
-		val->cell_len = __wt_cell_pack_ovfl(session, &val->cell,
-		    WT_CELL_VALUE_OVFL, start_ts, stop_ts,
-		    start_txn, stop_txn, rle, value->size);
-		val->buf.data = value->data;
-		val->buf.size = value->size;
-		val->len = val->cell_len + value->size;
-	} else
-		WT_RET(__rec_cell_build_val(session, r,
-		    value->data, value->size,
-		    start_ts, stop_ts, start_txn, stop_txn, rle));
-
-	/* Boundary: split or write the page. */
-	if (__rec_need_split(r, val->len))
-		WT_RET(__rec_split_crossing_bnd(session, r, val->len));
-
-	/* Copy the value onto the page. */
-	if (!deleted && !overflow_type && btree->dictionary)
-		WT_RET(__rec_dict_replace(session, r,
-		    start_ts, stop_ts, start_txn, stop_txn, rle, val));
-	__rec_image_copy(session, r, val);
-	__rec_addr_ts_update(r,
-	    start_ts, durable_ts, stop_ts, start_txn, stop_txn);
-
-	/* Update the starting record number in case we split. */
-	r->recno += rle;
-
-	return (0);
-}
-
-/*
- * __rec_col_var --
- *	Reconcile a variable-width column-store leaf page.
- */
-static int
-__rec_col_var(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *pageref, WT_SALVAGE_COOKIE *salvage)
-{
-	enum { OVFL_IGNORE, OVFL_UNUSED, OVFL_USED } ovfl_state;
-	struct {
-		WT_ITEM	*value;				/* Value */
-		wt_timestamp_t start_ts, stop_ts;	/* Timestamps */
-		uint64_t start_txn, stop_txn;		/* Transactions */
-		bool deleted;				/* If deleted */
-	} last;
-	WT_ADDR *addr;
-	WT_BTREE *btree;
-	WT_CELL *cell;
-	WT_CELL_UNPACK *vpack, _vpack;
-	WT_COL *cip;
-	WT_CURSOR_BTREE *cbt;
-	WT_DECL_ITEM(orig);
-	WT_DECL_RET;
-	WT_INSERT *ins;
-	WT_PAGE *page;
-	WT_UPDATE *upd;
-	WT_UPDATE_SELECT upd_select;
-	wt_timestamp_t start_ts, durable_ts, newest_durable_ts, stop_ts;
-	uint64_t n, nrepeat, repeat_count, rle, skip, src_recno;
-	uint64_t start_txn, stop_txn;
-	uint32_t i, size;
-	bool deleted, orig_deleted, update_no_copy;
-	const void *data;
-
-	btree = S2BT(session);
-	vpack = &_vpack;
-	cbt = &r->update_modify_cbt;
-	page = pageref->page;
-	upd = NULL;
-	size = 0;
-	data = NULL;
-
-	/*
-	 * Acquire the newest-durable timestamp for this page so we can roll it
-	 * forward. If it exists, it's in the WT_REF structure or the parent's
-	 * disk image.
-	 */
-	if ((addr = pageref->addr) == NULL)
-		newest_durable_ts = WT_TS_NONE;
-	else if (__wt_off_page(pageref->home, addr))
-		newest_durable_ts = addr->newest_durable_ts;
-	else {
-		__wt_cell_unpack(session, pageref->home, pageref->addr, vpack);
-		newest_durable_ts = vpack->newest_durable_ts;
-	}
-
-	/* Set the "last" values to cause failure if they're not set. */
-	last.value = r->last;
-	last.start_ts = WT_TS_MAX;
-	last.stop_ts = WT_TS_NONE;
-	last.start_txn = WT_TXN_MAX;
-	last.stop_txn = WT_TXN_NONE;
-	last.deleted = false;
-
-	/*
-	 * Set the start/stop values to cause failure if they're not set.
-	 * [-Werror=maybe-uninitialized]
-	 */
-	/* NOLINTNEXTLINE(clang-analyzer-deadcode.DeadStores) */
-	start_ts = WT_TS_MAX;
-	durable_ts = stop_ts = WT_TS_NONE;
-	start_txn = WT_TXN_MAX;
-	stop_txn = WT_TS_NONE;
-
-	WT_RET(__rec_split_init(session,
-	    r, page, pageref->ref_recno, btree->maxleafpage_precomp));
-
-	WT_RET(__wt_scr_alloc(session, 0, &orig));
-
-	/*
-	 * The salvage code may be calling us to reconcile a page where there
-	 * were missing records in the column-store name space.  If taking the
-	 * first record from on the page, it might be a deleted record, so we
-	 * have to give the RLE code a chance to figure that out.  Else, if
-	 * not taking the first record from the page, write a single element
-	 * representing the missing records onto a new page.  (Don't pass the
-	 * salvage cookie to our helper function in this case, we're handling
-	 * one of the salvage cookie fields on our own, and we don't need the
-	 * helper function's assistance.)
-	 */
-	rle = 0;
-	if (salvage != NULL && salvage->missing != 0) {
-		if (salvage->skip == 0) {
-			rle = salvage->missing;
-			last.start_ts = WT_TS_NONE;
-			last.stop_ts = WT_TS_MAX;
-			last.start_txn = WT_TXN_NONE;
-			last.stop_txn = WT_TXN_MAX;
-			last.deleted = true;
-
-			/*
-			 * Correct the number of records we're going to "take",
-			 * pretending the missing records were on the page.
-			 */
-			salvage->take += salvage->missing;
-		} else
-			WT_ERR(__rec_col_var_helper(session, r, NULL, NULL,
-			    WT_TS_NONE, WT_TS_NONE, WT_TS_MAX,
-			    WT_TXN_NONE, WT_TXN_MAX,
-			    salvage->missing, true, false));
-	}
-
-	/*
-	 * We track two data items through this loop: the previous (last) item
-	 * and the current item: if the last item is the same as the current
-	 * item, we increment the RLE count for the last item; if the last item
-	 * is different from the current item, we write the last item onto the
-	 * page, and replace it with the current item.  The r->recno counter
-	 * tracks records written to the page, and is incremented by the helper
-	 * function immediately after writing records to the page.  The record
-	 * number of our source record, that is, the current item, is maintained
-	 * in src_recno.
-	 */
-	src_recno = r->recno + rle;
-
-	/* For each entry in the in-memory page... */
-	WT_COL_FOREACH(page, cip, i) {
-		ovfl_state = OVFL_IGNORE;
-		cell = WT_COL_PTR(page, cip);
-		__wt_cell_unpack(session, page, cell, vpack);
-		nrepeat = __wt_cell_rle(vpack);
-		ins = WT_SKIP_FIRST(WT_COL_UPDATE(page, cip));
-
-		/*
-		 * If the original value is "deleted", there's no value
-		 * to compare, we're done.
-		 */
-		orig_deleted = vpack->type == WT_CELL_DEL;
-		if (orig_deleted)
-			goto record_loop;
-
-		/*
-		 * Overflow items are tricky: we don't know until we're
-		 * finished processing the set of values if we need the
-		 * overflow value or not.  If we don't use the overflow
-		 * item at all, we have to discard it from the backing
-		 * file, otherwise we'll leak blocks on the checkpoint.
-		 * That's safe because if the backing overflow value is
-		 * still needed by any running transaction, we'll cache
-		 * a copy in the update list.
-		 *
-		 * Regardless, we avoid copying in overflow records: if
-		 * there's a WT_INSERT entry that modifies a reference
-		 * counted overflow record, we may have to write copies
-		 * of the overflow record, and in that case we'll do the
-		 * comparisons, but we don't read overflow items just to
-		 * see if they match records on either side.
-		 */
-		if (vpack->ovfl) {
-			ovfl_state = OVFL_UNUSED;
-			goto record_loop;
-		}
-
-		/*
-		 * If data is Huffman encoded, we have to decode it in
-		 * order to compare it with the last item we saw, which
-		 * may have been an update string.  This guarantees we
-		 * find every single pair of objects we can RLE encode,
-		 * including applications updating an existing record
-		 * where the new value happens (?) to match a Huffman-
-		 * encoded value in a previous or next record.
-		 */
-		WT_ERR(__wt_dsk_cell_data_ref(
-		    session, WT_PAGE_COL_VAR, vpack, orig));
-
-record_loop:	/*
-		 * Generate on-page entries: loop repeat records, looking for
-		 * WT_INSERT entries matching the record number.  The WT_INSERT
-		 * lists are in sorted order, so only need check the next one.
-		 */
-		for (n = 0;
-		    n < nrepeat; n += repeat_count, src_recno += repeat_count) {
-			start_ts = vpack->start_ts;
-			durable_ts = newest_durable_ts;
-			stop_ts = vpack->stop_ts;
-			start_txn = vpack->start_txn;
-			stop_txn = vpack->stop_txn;
-			upd = NULL;
-			if (ins != NULL && WT_INSERT_RECNO(ins) == src_recno) {
-				WT_ERR(__wt_rec_upd_select(
-				    session, r, ins, cip, vpack, &upd_select));
-				upd = upd_select.upd;
-				if (upd == NULL) {
-					/*
-					 * TIMESTAMP-FIXME
-					 * I'm pretty sure this is wrong: a NULL
-					 * update means an item was deleted, and
-					 * I think that requires a tombstone on
-					 * the page.
-					 */
-					start_ts = WT_TS_NONE;
-					durable_ts = WT_TS_NONE;
-					stop_ts = WT_TS_MAX;
-					start_txn = WT_TXN_NONE;
-					stop_txn = WT_TXN_MAX;
-				} else {
-					start_ts = upd_select.start_ts;
-					durable_ts = upd_select.durable_ts;
-					stop_ts = upd_select.stop_ts;
-					start_txn = upd_select.start_txn;
-					stop_txn = upd_select.stop_txn;
-				}
-				ins = WT_SKIP_NEXT(ins);
-			}
-
-			update_no_copy = true;	/* No data copy */
-			repeat_count = 1;	/* Single record */
-			deleted = false;
-
-			if (upd != NULL) {
-				switch (upd->type) {
-				case WT_UPDATE_MODIFY:
-					cbt->slot = WT_COL_SLOT(page, cip);
-					WT_ERR(__wt_value_return_upd(
-					    session, cbt, upd,
-					    F_ISSET(r, WT_REC_VISIBLE_ALL)));
-					data = cbt->iface.value.data;
-					size = (uint32_t)cbt->iface.value.size;
-					update_no_copy = false;
-					break;
-				case WT_UPDATE_STANDARD:
-					data = upd->data;
-					size = upd->size;
-					break;
-				case WT_UPDATE_TOMBSTONE:
-					deleted = true;
-					break;
-				WT_ILLEGAL_VALUE_ERR(session, upd->type);
-				}
-			} else if (vpack->raw == WT_CELL_VALUE_OVFL_RM) {
-				/*
-				 * If doing an update save and restore, and the
-				 * underlying value is a removed overflow value,
-				 * we end up here.
-				 *
-				 * If necessary, when the overflow value was
-				 * originally removed, reconciliation appended
-				 * a globally visible copy of the value to the
-				 * key's update list, meaning the on-page item
-				 * isn't accessed after page re-instantiation.
-				 *
-				 * Assert the case.
-				 */
-				WT_ASSERT(session,
-				    F_ISSET(r, WT_REC_UPDATE_RESTORE));
-
-				/*
-				 * The on-page value will never be accessed,
-				 * write a placeholder record.
-				 */
-				data = "ovfl-unused";
-				size = WT_STORE_SIZE(strlen("ovfl-unused"));
-			} else {
-				update_no_copy = false;	/* Maybe data copy */
-
-				/*
-				 * The repeat count is the number of records up
-				 * to the next WT_INSERT record, or up to the
-				 * end of the entry if we have no more WT_INSERT
-				 * records.
-				 */
-				if (ins == NULL)
-					repeat_count = nrepeat - n;
-				else
-					repeat_count =
-					    WT_INSERT_RECNO(ins) - src_recno;
-
-				deleted = orig_deleted;
-				if (deleted)
-					goto compare;
-
-				/*
-				 * If we are handling overflow items, use the
-				 * overflow item itself exactly once, after
-				 * which we have to copy it into a buffer and
-				 * from then on use a complete copy because we
-				 * are re-creating a new overflow record each
-				 * time.
-				 */
-				switch (ovfl_state) {
-				case OVFL_UNUSED:
-					/*
-					 * An as-yet-unused overflow item.
-					 *
-					 * We're going to copy the on-page cell,
-					 * write out any record we're tracking.
-					 */
-					if (rle != 0) {
-						WT_ERR(__rec_col_var_helper(
-						    session, r, salvage,
-						    last.value, last.start_ts,
-						    durable_ts, last.stop_ts,
-						    last.start_txn,
-						    last.stop_txn,
-						    rle, last.deleted, false));
-						rle = 0;
-					}
-
-					last.value->data = vpack->data;
-					last.value->size = vpack->size;
-					WT_ERR(__rec_col_var_helper(session, r,
-					    salvage, last.value,
-					    start_ts, durable_ts, stop_ts,
-					    start_txn, stop_txn,
-					    repeat_count, false, true));
-
-					/* Track if page has overflow items. */
-					r->ovfl_items = true;
-
-					ovfl_state = OVFL_USED;
-					continue;
-				case OVFL_USED:
-					/*
-					 * Original is an overflow item; we used
-					 * it for a key and now we need another
-					 * copy; read it into memory.
-					 */
-					WT_ERR(__wt_dsk_cell_data_ref(session,
-					    WT_PAGE_COL_VAR, vpack, orig));
-
-					ovfl_state = OVFL_IGNORE;
-					/* FALLTHROUGH */
-				case OVFL_IGNORE:
-					/*
-					 * Original is an overflow item and we
-					 * were forced to copy it into memory,
-					 * or the original wasn't an overflow
-					 * item; use the data copied into orig.
-					 */
-					data = orig->data;
-					size = (uint32_t)orig->size;
-					break;
-				}
-			}
-
-compare:		/*
-			 * If we have a record against which to compare, and
-			 * the records compare equal, increment the rle counter
-			 * and continue.  If the records don't compare equal,
-			 * output the last record and swap the last and current
-			 * buffers: do NOT update the starting record number,
-			 * we've been doing that all along.
-			 */
-			if (rle != 0) {
-				if ((!__wt_process.page_version_ts ||
-				    (last.start_ts == start_ts &&
-				    last.stop_ts == stop_ts &&
-				    last.start_txn == start_txn &&
-				    last.stop_txn == stop_txn)) &&
-				    ((deleted && last.deleted) ||
-				    (!deleted && !last.deleted &&
-				    last.value->size == size &&
-				    memcmp(
-				    last.value->data, data, size) == 0))) {
-					rle += repeat_count;
-					continue;
-				}
-				WT_ERR(__rec_col_var_helper(session, r, salvage,
-				    last.value,
-				    last.start_ts, durable_ts, last.stop_ts,
-				    last.start_txn, last.stop_txn,
-				    rle, last.deleted, false));
-			}
-
-			/*
-			 * Swap the current/last state.
-			 *
-			 * Reset RLE counter and turn on comparisons.
-			 */
-			if (!deleted) {
-				/*
-				 * We can't simply assign the data values into
-				 * the last buffer because they may have come
-				 * from a copy built from an encoded/overflow
-				 * cell and creating the next record is going
-				 * to overwrite that memory.  Check, because
-				 * encoded/overflow cells aren't that common
-				 * and we'd like to avoid the copy.  If data
-				 * was taken from the current unpack structure
-				 * (which points into the page), or was taken
-				 * from an update structure, we can just use
-				 * the pointers, they're not moving.
-				 */
-				if (data == vpack->data || update_no_copy) {
-					last.value->data = data;
-					last.value->size = size;
-				} else
-					WT_ERR(__wt_buf_set(
-					    session, last.value, data, size));
-			}
-			last.start_ts = start_ts;
-			last.stop_ts = stop_ts;
-			last.start_txn = start_txn;
-			last.stop_txn = stop_txn;
-			last.deleted = deleted;
-			rle = repeat_count;
-		}
-
-		/*
-		 * The first time we find an overflow record we never used,
-		 * discard the underlying blocks, they're no longer useful.
-		 */
-		if (ovfl_state == OVFL_UNUSED &&
-		    vpack->raw != WT_CELL_VALUE_OVFL_RM)
-			WT_ERR(__wt_ovfl_remove(
-			    session, page, vpack, F_ISSET(r, WT_REC_EVICT)));
-	}
-
-	/* Walk any append list. */
-	for (ins =
-	    WT_SKIP_FIRST(WT_COL_APPEND(page));; ins = WT_SKIP_NEXT(ins)) {
-		if (ins == NULL) {
-			/*
-			 * If the page split, instantiate any missing records in
-			 * the page's name space. (Imagine record 98 is
-			 * transactionally visible, 99 wasn't created or is not
-			 * yet visible, 100 is visible. Then the page splits and
-			 * record 100 moves to another page. When we reconcile
-			 * the original page, we write record 98, then we don't
-			 * see record 99 for whatever reason. If we've moved
-			 * record 100, we don't know to write a deleted record
-			 * 99 on the page.)
-			 *
-			 * Assert the recorded record number is past the end of
-			 * the page.
-			 *
-			 * The record number recorded during the split is the
-			 * first key on the split page, that is, one larger than
-			 * the last key on this page, we have to decrement it.
-			 */
-			if ((n = page->
-			    modify->mod_col_split_recno) == WT_RECNO_OOB)
-				break;
-			WT_ASSERT(session, n >= src_recno);
-			n -= 1;
-
-			upd = NULL;
-		} else {
-			WT_ERR(__wt_rec_upd_select(
-			    session, r, ins, NULL, NULL, &upd_select));
-			upd = upd_select.upd;
-			n = WT_INSERT_RECNO(ins);
-		}
-		if (upd == NULL) {
-			/*
-			 * TIMESTAMP-FIXME
-			 * I'm pretty sure this is wrong: a NULL update means
-			 * an item was deleted, and I think that requires a
-			 * tombstone on the page.
-			 */
-			start_ts = WT_TS_NONE;
-			durable_ts = WT_TS_NONE;
-			stop_ts = WT_TS_MAX;
-			start_txn = WT_TXN_NONE;
-			stop_txn = WT_TXN_MAX;
-		} else {
-			start_ts = upd_select.start_ts;
-			durable_ts = upd_select.durable_ts;
-			stop_ts = upd_select.stop_ts;
-			start_txn = upd_select.start_txn;
-			stop_txn = upd_select.stop_txn;
-		}
-		while (src_recno <= n) {
-			deleted = false;
-			update_no_copy = true;
-
-			/*
-			 * The application may have inserted records which left
-			 * gaps in the name space, and these gaps can be huge.
-			 * If we're in a set of deleted records, skip the boring
-			 * part.
-			 */
-			if (src_recno < n) {
-				deleted = true;
-				if (last.deleted &&
-				    (!__wt_process.page_version_ts ||
-				    (last.start_ts == start_ts &&
-				    last.stop_ts == stop_ts &&
-				    last.start_txn == start_txn &&
-				    last.stop_txn == stop_txn))) {
-					/*
-					 * The record adjustment is decremented
-					 * by one so we can naturally fall into
-					 * the RLE accounting below, where we
-					 * increment rle by one, then continue
-					 * in the outer loop, where we increment
-					 * src_recno by one.
-					 */
-					skip = (n - src_recno) - 1;
-					rle += skip;
-					src_recno += skip;
-				}
-			} else if (upd == NULL) {
-				/*
-				 * TIMESTAMP-FIXME
-				 * I'm pretty sure this is wrong: a NULL
-				 * update means an item was deleted, and
-				 * I think that requires a tombstone on
-				 * the page.
-				 */
-				start_ts = WT_TS_NONE;
-				durable_ts = WT_TS_NONE;
-				stop_ts = WT_TS_MAX;
-				start_txn = WT_TXN_NONE;
-				stop_txn = WT_TXN_MAX;
-
-				deleted = true;
-			} else {
-				start_ts = upd_select.start_ts;
-				durable_ts = upd_select.durable_ts;
-				stop_ts = upd_select.stop_ts;
-				start_txn = upd_select.start_txn;
-				stop_txn = upd_select.stop_txn;
-
-				switch (upd->type) {
-				case WT_UPDATE_MODIFY:
-					/*
-					 * Impossible slot, there's no backing
-					 * on-page item.
-					 */
-					cbt->slot = UINT32_MAX;
-					WT_ERR(__wt_value_return_upd(
-					    session, cbt, upd,
-					    F_ISSET(r, WT_REC_VISIBLE_ALL)));
-					data = cbt->iface.value.data;
-					size = (uint32_t)cbt->iface.value.size;
-					update_no_copy = false;
-					break;
-				case WT_UPDATE_STANDARD:
-					data = upd->data;
-					size = upd->size;
-					break;
-				case WT_UPDATE_TOMBSTONE:
-					deleted = true;
-					break;
-				WT_ILLEGAL_VALUE_ERR(session, upd->type);
-				}
-			}
-
-			/*
-			 * Handle RLE accounting and comparisons -- see comment
-			 * above, this code fragment does the same thing.
-			 */
-			if (rle != 0) {
-				if ((!__wt_process.page_version_ts ||
-				    (last.start_ts == start_ts &&
-				    last.stop_ts == stop_ts &&
-				    last.start_txn == start_txn &&
-				    last.stop_txn == stop_txn)) &&
-				    ((deleted && last.deleted) ||
-				    (!deleted && !last.deleted &&
-				    last.value->size == size &&
-				    memcmp(
-				    last.value->data, data, size) == 0))) {
-					++rle;
-					goto next;
-				}
-				WT_ERR(__rec_col_var_helper(session, r, salvage,
-				    last.value,
-				    last.start_ts, durable_ts, last.stop_ts,
-				    last.start_txn, last.stop_txn,
-				    rle, last.deleted, false));
-			}
-
-			/*
-			 * Swap the current/last state. We can't simply assign
-			 * the data values into the last buffer because they may
-			 * be a temporary copy built from a chain of modified
-			 * updates and creating the next record will overwrite
-			 * that memory. Check, we'd like to avoid the copy. If
-			 * data was taken from an update structure, we can just
-			 * use the pointers, they're not moving.
-			 */
-			if (!deleted) {
-				if (update_no_copy) {
-					last.value->data = data;
-					last.value->size = size;
-				} else
-					WT_ERR(__wt_buf_set(
-					    session, last.value, data, size));
-			}
-
-			/* Ready for the next loop, reset the RLE counter. */
-			last.start_ts = start_ts;
-			last.stop_ts = stop_ts;
-			last.start_txn = start_txn;
-			last.stop_txn = stop_txn;
-			last.deleted = deleted;
-			rle = 1;
-
-			/*
-			 * Move to the next record. It's not a simple increment
-			 * because if it's the maximum record, incrementing it
-			 * wraps to 0 and this turns into an infinite loop.
-			 */
-next:			if (src_recno == UINT64_MAX)
-				break;
-			++src_recno;
-		}
-
-		/*
-		 * Execute this loop once without an insert item to catch any
-		 * missing records due to a split, then quit.
-		 */
-		if (ins == NULL)
-			break;
-	}
-
-	/* If we were tracking a record, write it. */
-	if (rle != 0)
-		WT_ERR(__rec_col_var_helper(session, r, salvage,
-		    last.value, last.start_ts, durable_ts, last.stop_ts,
-		    last.start_txn, last.stop_txn, rle, last.deleted, false));
-
-	/* Write the remnant page. */
-	ret = __rec_split_finish(session, r);
-
-err:	__wt_scr_free(session, &orig);
-	return (ret);
-}
-
-/*
- * __rec_row_int --
- *	Reconcile a row-store internal page.
- */
-static int
-__rec_row_int(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_PAGE *page)
-{
-	WT_ADDR *addr;
-	WT_BTREE *btree;
-	WT_CELL *cell;
-	WT_CELL_UNPACK *kpack, _kpack, *vpack, _vpack;
-	WT_CHILD_STATE state;
-	WT_DECL_RET;
-	WT_IKEY *ikey;
-	WT_PAGE *child;
-	WT_REC_KV *key, *val;
-	WT_REF *ref;
-	wt_timestamp_t oldest_start_ts, newest_durable_ts, newest_stop_ts;
-	size_t size;
-	uint64_t oldest_start_txn, newest_stop_txn;
-	bool hazard, key_onpage_ovfl, ovfl_key;
-	const void *p;
-
-	btree = S2BT(session);
-	child = NULL;
-	hazard = false;
-
-	key = &r->k;
-	kpack = &_kpack;
-	WT_CLEAR(*kpack);	/* -Wuninitialized */
-	val = &r->v;
-	vpack = &_vpack;
-	WT_CLEAR(*vpack);	/* -Wuninitialized */
-
-	ikey = NULL;		/* -Wuninitialized */
-	cell = NULL;
-	key_onpage_ovfl = false;
-
-	WT_RET(__rec_split_init(
-	    session, r, page, 0, btree->maxintlpage_precomp));
-
-	/*
-	 * Ideally, we'd never store the 0th key on row-store internal pages
-	 * because it's never used during tree search and there's no reason
-	 * to waste the space.  The problem is how we do splits: when we split,
-	 * we've potentially picked out several "split points" in the buffer
-	 * which is overflowing the maximum page size, and when the overflow
-	 * happens, we go back and physically split the buffer, at those split
-	 * points, into new pages.  It would be both difficult and expensive
-	 * to re-process the 0th key at each split point to be an empty key,
-	 * so we don't do that.  However, we are reconciling an internal page
-	 * for whatever reason, and the 0th key is known to be useless.  We
-	 * truncate the key to a single byte, instead of removing it entirely,
-	 * it simplifies various things in other parts of the code (we don't
-	 * have to special case transforming the page from its disk image to
-	 * its in-memory version, for example).
-	 */
-	r->cell_zero = true;
-
-	/* For each entry in the in-memory page... */
-	WT_INTL_FOREACH_BEGIN(session, page, ref) {
-		/*
-		 * There are different paths if the key is an overflow item vs.
-		 * a straight-forward on-page value. If an overflow item, we
-		 * would have instantiated it, and we can use that fact to set
-		 * things up.
-		 *
-		 * Note the cell reference and unpacked key cell are available
-		 * only in the case of an instantiated, off-page key, we don't
-		 * bother setting them if that's not possible.
-		 */
-		if (F_ISSET_ATOMIC(page, WT_PAGE_OVERFLOW_KEYS)) {
-			cell = NULL;
-			key_onpage_ovfl = false;
-			ikey = __wt_ref_key_instantiated(ref);
-			if (ikey != NULL && ikey->cell_offset != 0) {
-				cell =
-				    WT_PAGE_REF_OFFSET(page, ikey->cell_offset);
-				__wt_cell_unpack(session, page, cell, kpack);
-				key_onpage_ovfl = kpack->ovfl &&
-				    kpack->raw != WT_CELL_KEY_OVFL_RM;
-			}
-		}
-
-		WT_ERR(__rec_child_modify(session, r, ref, &hazard, &state));
-		addr = ref->addr;
-		child = ref->page;
-
-		switch (state) {
-		case WT_CHILD_IGNORE:
-			/*
-			 * Ignored child.
-			 *
-			 * Overflow keys referencing pages we're not writing are
-			 * no longer useful, schedule them for discard.  Don't
-			 * worry about instantiation, internal page keys are
-			 * always instantiated.  Don't worry about reuse,
-			 * reusing this key in this reconciliation is unlikely.
-			 */
-			if (key_onpage_ovfl)
-				WT_ERR(__wt_ovfl_discard_add(
-				    session, page, kpack->cell));
-			WT_CHILD_RELEASE_ERR(session, hazard, ref);
-			continue;
-
-		case WT_CHILD_MODIFIED:
-			/*
-			 * Modified child.  Empty pages are merged into the
-			 * parent and discarded.
-			 */
-			switch (child->modify->rec_result) {
-			case WT_PM_REC_EMPTY:
-				/*
-				 * Overflow keys referencing empty pages are no
-				 * longer useful, schedule them for discard.
-				 * Don't worry about instantiation, internal
-				 * page keys are always instantiated.  Don't
-				 * worry about reuse, reusing this key in this
-				 * reconciliation is unlikely.
-				 */
-				if (key_onpage_ovfl)
-					WT_ERR(__wt_ovfl_discard_add(
-					    session, page, kpack->cell));
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_MULTIBLOCK:
-				/*
-				 * Overflow keys referencing split pages are no
-				 * longer useful (the split page's key is the
-				 * interesting key); schedule them for discard.
-				 * Don't worry about instantiation, internal
-				 * page keys are always instantiated.  Don't
-				 * worry about reuse, reusing this key in this
-				 * reconciliation is unlikely.
-				 */
-				if (key_onpage_ovfl)
-					WT_ERR(__wt_ovfl_discard_add(
-					    session, page, kpack->cell));
-
-				WT_ERR(__rec_row_merge(session, r, child));
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_REPLACE:
-				/*
-				 * If the page is replaced, the page's modify
-				 * structure has the page's address.
-				 */
-				addr = &child->modify->mod_replace;
-				break;
-			WT_ILLEGAL_VALUE_ERR(
-			    session, child->modify->rec_result);
-			}
-			break;
-		case WT_CHILD_ORIGINAL:
-			/* Original child. */
-			break;
-		case WT_CHILD_PROXY:
-			/* Deleted child where we write a proxy cell. */
-			break;
-		}
-
-		/*
-		 * Build the value cell, the child page's address.  Addr points
-		 * to an on-page cell or an off-page WT_ADDR structure. There's
-		 * a special cell type in the case of page deletion requiring
-		 * a proxy cell, otherwise use the information from the addr or
-		 * original cell.
-		 */
-		if (__wt_off_page(page, addr)) {
-			__rec_cell_build_addr(session, r, addr,
-			    state == WT_CHILD_PROXY, WT_RECNO_OOB);
-			oldest_start_ts = addr->oldest_start_ts;
-			newest_durable_ts = addr->newest_durable_ts;
-			newest_stop_ts = addr->newest_stop_ts;
-			oldest_start_txn = addr->oldest_start_txn;
-			newest_stop_txn = addr->newest_stop_txn;
-		} else {
-			__wt_cell_unpack(session, page, ref->addr, vpack);
-			if (state == WT_CHILD_PROXY) {
-				WT_ERR(__wt_buf_set(session, &val->buf,
-				    ref->addr, __wt_cell_total_len(vpack)));
-				__wt_cell_type_reset(session,
-				    val->buf.mem, 0, WT_CELL_ADDR_DEL);
-			} else {
-				val->buf.data = ref->addr;
-				val->buf.size = __wt_cell_total_len(vpack);
-			}
-			val->cell_len = 0;
-			val->len = val->buf.size;
-			oldest_start_ts = vpack->oldest_start_ts;
-			newest_durable_ts = vpack->newest_durable_ts;
-			newest_stop_ts = vpack->newest_stop_ts;
-			oldest_start_txn = vpack->oldest_start_txn;
-			newest_stop_txn = vpack->newest_stop_txn;
-		}
-		WT_CHILD_RELEASE_ERR(session, hazard, ref);
-
-		/*
-		 * Build key cell.
-		 * Truncate any 0th key, internal pages don't need 0th keys.
-		 */
-		if (key_onpage_ovfl) {
-			key->buf.data = cell;
-			key->buf.size = __wt_cell_total_len(kpack);
-			key->cell_len = 0;
-			key->len = key->buf.size;
-			ovfl_key = true;
-		} else {
-			__wt_ref_key(page, ref, &p, &size);
-			WT_ERR(__rec_cell_build_int_key(
-			    session, r, p, r->cell_zero ? 1 : size, &ovfl_key));
-		}
-		r->cell_zero = false;
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len)) {
-			/*
-			 * In one path above, we copied address blocks from the
-			 * page rather than building the actual key. In that
-			 * case, we have to build the key now because we are
-			 * about to promote it.
-			 */
-			if (key_onpage_ovfl) {
-				WT_ERR(__wt_buf_set(session, r->cur,
-				    WT_IKEY_DATA(ikey), ikey->size));
-				key_onpage_ovfl = false;
-			}
-
-			WT_ERR(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-		}
-
-		/* Copy the key and value onto the page. */
-		__rec_image_copy(session, r, key);
-		__rec_image_copy(session, r, val);
-		__rec_addr_ts_update(r,
-		    oldest_start_ts, newest_durable_ts, newest_stop_ts,
-		    oldest_start_txn, newest_stop_txn);
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-	} WT_INTL_FOREACH_END;
-
-	/* Write the remnant page. */
-	return (__rec_split_finish(session, r));
-
-err:	WT_CHILD_RELEASE(session, hazard, ref);
-	return (ret);
-}
-
-/*
- * __rec_row_merge --
- *	Merge in a split page.
- */
-static int
-__rec_row_merge(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_PAGE *page)
-{
-	WT_ADDR *addr;
-	WT_MULTI *multi;
-	WT_PAGE_MODIFY *mod;
-	WT_REC_KV *key, *val;
-	uint32_t i;
-	bool ovfl_key;
-
-	mod = page->modify;
-
-	key = &r->k;
-	val = &r->v;
-
-	/* For each entry in the split array... */
-	for (multi = mod->mod_multi,
-	    i = 0; i < mod->mod_multi_entries; ++multi, ++i) {
-		/* Build the key and value cells. */
-		WT_RET(__rec_cell_build_int_key(session, r,
-		    WT_IKEY_DATA(multi->key.ikey),
-		    r->cell_zero ? 1 : multi->key.ikey->size, &ovfl_key));
-		r->cell_zero = false;
-
-		addr = &multi->addr;
-		__rec_cell_build_addr(session, r, addr, false, WT_RECNO_OOB);
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len))
-			WT_RET(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-
-		/* Copy the key and value onto the page. */
-		__rec_image_copy(session, r, key);
-		__rec_image_copy(session, r, val);
-		__rec_addr_ts_update(r, addr->oldest_start_ts,
-		    addr->newest_durable_ts, addr->newest_stop_ts,
-		    addr->oldest_start_txn, addr->newest_stop_txn);
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-	}
-	return (0);
-}
-
-/*
- * __rec_row_zero_len --
- *	Return if a zero-length item can be written.
- */
-static bool
-__rec_row_zero_len(WT_SESSION_IMPL *session,
-    wt_timestamp_t start_ts, wt_timestamp_t stop_ts,
-    uint64_t start_txn, uint64_t stop_txn)
-{
-	/* Before timestamps were stored on pages, it was always possible. */
-	if (!__wt_process.page_version_ts)
-		return (true);
-
-	/*
-	 * The item must be globally visible because we're not writing anything
-	 * on the page.
-	 */
-	return ((stop_ts == WT_TS_MAX && stop_txn == WT_TXN_MAX) &&
-	    ((start_ts == WT_TS_NONE && start_txn == WT_TXN_NONE) ||
-	    __wt_txn_visible_all(session, start_txn, start_ts)));
-}
-
-/*
- * __rec_row_leaf --
- *	Reconcile a row-store leaf page.
- */
-static int
-__rec_row_leaf(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *pageref, WT_SALVAGE_COOKIE *salvage)
-{
-	WT_ADDR *addr;
-	WT_BTREE *btree;
-	WT_CELL *cell;
-	WT_CELL_UNPACK *kpack, _kpack, *vpack, _vpack;
-	WT_CURSOR_BTREE *cbt;
-	WT_DECL_ITEM(tmpkey);
-	WT_DECL_ITEM(tmpval);
-	WT_DECL_RET;
-	WT_IKEY *ikey;
-	WT_INSERT *ins;
-	WT_PAGE *page;
-	WT_REC_KV *key, *val;
-	WT_ROW *rip;
-	WT_UPDATE *upd;
-	WT_UPDATE_SELECT upd_select;
-	wt_timestamp_t start_ts, durable_ts, newest_durable_ts, stop_ts;
-	size_t size;
-	uint64_t slvg_skip, start_txn, stop_txn;
-	uint32_t i;
-	bool dictionary, key_onpage_ovfl, ovfl_key;
-	void *copy;
-	const void *p;
-
-	btree = S2BT(session);
-	cbt = &r->update_modify_cbt;
-	page = pageref->page;
-	slvg_skip = salvage == NULL ? 0 : salvage->skip;
-
-	key = &r->k;
-	val = &r->v;
-	vpack = &_vpack;
-
-	/*
-	 * Acquire the newest-durable timestamp for this page so we can roll it
-	 * forward. If it exists, it's in the WT_REF structure or the parent's
-	 * disk image.
-	 */
-	if ((addr = pageref->addr) == NULL)
-		newest_durable_ts = WT_TS_NONE;
-	else if (__wt_off_page(pageref->home, addr))
-		newest_durable_ts = addr->newest_durable_ts;
-	else {
-		__wt_cell_unpack(session, pageref->home, pageref->addr, vpack);
-		newest_durable_ts = vpack->newest_durable_ts;
-	}
-
-	WT_RET(__rec_split_init(
-	    session, r, page, 0, btree->maxleafpage_precomp));
-
-	/*
-	 * Write any K/V pairs inserted into the page before the first from-disk
-	 * key on the page.
-	 */
-	if ((ins = WT_SKIP_FIRST(WT_ROW_INSERT_SMALLEST(page))) != NULL)
-		WT_RET(__rec_row_leaf_insert(session, r, ins));
-
-	/*
-	 * Temporary buffers in which to instantiate any uninstantiated keys
-	 * or value items we need.
-	 */
-	WT_ERR(__wt_scr_alloc(session, 0, &tmpkey));
-	WT_ERR(__wt_scr_alloc(session, 0, &tmpval));
-
-	/* For each entry in the page... */
-	WT_ROW_FOREACH(page, rip, i) {
-		/*
-		 * The salvage code, on some rare occasions, wants to reconcile
-		 * a page but skip some leading records on the page.  Because
-		 * the row-store leaf reconciliation function copies keys from
-		 * the original disk page, this is non-trivial -- just changing
-		 * the in-memory pointers isn't sufficient, we have to change
-		 * the WT_CELL structures on the disk page, too.  It's ugly, but
-		 * we pass in a value that tells us how many records to skip in
-		 * this case.
-		 */
-		if (slvg_skip != 0) {
-			--slvg_skip;
-			continue;
-		}
-
-		/*
-		 * Figure out the key: set any cell reference (and unpack it),
-		 * set any instantiated key reference.
-		 */
-		copy = WT_ROW_KEY_COPY(rip);
-		(void)__wt_row_leaf_key_info(
-		    page, copy, &ikey, &cell, NULL, NULL);
-		if (cell == NULL)
-			kpack = NULL;
-		else {
-			kpack = &_kpack;
-			__wt_cell_unpack(session, page, cell, kpack);
-		}
-
-		/* Unpack the on-page value cell, set the default timestamps. */
-		__wt_row_leaf_value_cell(session, page, rip, NULL, vpack);
-		start_ts = vpack->start_ts;
-		durable_ts = newest_durable_ts;
-		stop_ts = vpack->stop_ts;
-		start_txn = vpack->start_txn;
-		stop_txn = vpack->stop_txn;
-
-		/* Look for an update. */
-		WT_ERR(__wt_rec_upd_select(
-		    session, r, NULL, rip, vpack, &upd_select));
-		if ((upd = upd_select.upd) != NULL) {
-			start_ts = upd_select.start_ts;
-			durable_ts = upd_select.durable_ts;
-			stop_ts = upd_select.stop_ts;
-			start_txn = upd_select.start_txn;
-			stop_txn = upd_select.stop_txn;
-		}
-
-		/* Build value cell. */
-		dictionary = false;
-		if (upd == NULL) {
-			/*
-			 * When the page was read into memory, there may not
-			 * have been a value item.
-			 *
-			 * If there was a value item, check if it's a dictionary
-			 * cell (a copy of another item on the page).  If it's a
-			 * copy, we have to create a new value item as the old
-			 * item might have been discarded from the page.
-			 */
-			if (vpack->raw == WT_CELL_VALUE_COPY) {
-				/* If the item is Huffman encoded, decode it. */
-				if (btree->huffman_value == NULL) {
-					p = vpack->data;
-					size = vpack->size;
-				} else {
-					WT_ERR(__wt_huffman_decode(session,
-					    btree->huffman_value,
-					    vpack->data, vpack->size,
-					    tmpval));
-					p = tmpval->data;
-					size = tmpval->size;
-				}
-				WT_ERR(__rec_cell_build_val(session, r, p, size,
-				    start_ts, stop_ts, start_txn, stop_txn, 0));
-				dictionary = true;
-			} else if (vpack->raw == WT_CELL_VALUE_OVFL_RM) {
-				/*
-				 * If doing an update save and restore, and the
-				 * underlying value is a removed overflow value,
-				 * we end up here.
-				 *
-				 * If necessary, when the overflow value was
-				 * originally removed, reconciliation appended
-				 * a globally visible copy of the value to the
-				 * key's update list, meaning the on-page item
-				 * isn't accessed after page re-instantiation.
-				 *
-				 * Assert the case.
-				 */
-				WT_ASSERT(session,
-				    F_ISSET(r, WT_REC_UPDATE_RESTORE));
-
-				/*
-				 * If the key is also a removed overflow item,
-				 * don't write anything at all.
-				 *
-				 * We don't have to write anything because the
-				 * code re-instantiating the page gets the key
-				 * to match the saved list of updates from the
-				 * original page.  By not putting the key on
-				 * the page, we'll move the key/value set from
-				 * a row-store leaf page slot to an insert list,
-				 * but that shouldn't matter.
-				 *
-				 * The reason we bother with the test is because
-				 * overflows are expensive to write.  It's hard
-				 * to imagine a real workload where this test is
-				 * worth the effort, but it's a simple test.
-				 */
-				if (kpack != NULL &&
-				    kpack->raw == WT_CELL_KEY_OVFL_RM)
-					goto leaf_insert;
-
-				/*
-				 * The on-page value will never be accessed,
-				 * write a placeholder record.
-				 */
-				WT_ERR(__rec_cell_build_val(session, r,
-				    "ovfl-unused", strlen("ovfl-unused"),
-				    start_ts, stop_ts, start_txn, stop_txn, 0));
-			} else {
-				val->buf.data = vpack->cell;
-				val->buf.size = __wt_cell_total_len(vpack);
-				val->cell_len = 0;
-				val->len = val->buf.size;
-
-				/* Track if page has overflow items. */
-				if (vpack->ovfl)
-					r->ovfl_items = true;
-			}
-		} else {
-			/*
-			 * The first time we find an overflow record we're not
-			 * going to use, discard the underlying blocks.
-			 */
-			if (vpack->ovfl && vpack->raw != WT_CELL_VALUE_OVFL_RM)
-				WT_ERR(__wt_ovfl_remove(session,
-				    page, vpack, F_ISSET(r, WT_REC_EVICT)));
-
-			switch (upd->type) {
-			case WT_UPDATE_MODIFY:
-				cbt->slot = WT_ROW_SLOT(page, rip);
-				WT_ERR(__wt_value_return_upd(session, cbt, upd,
-				    F_ISSET(r, WT_REC_VISIBLE_ALL)));
-				WT_ERR(__rec_cell_build_val(session, r,
-				    cbt->iface.value.data,
-				    cbt->iface.value.size,
-				    start_ts, stop_ts, start_txn, stop_txn, 0));
-				dictionary = true;
-				break;
-			case WT_UPDATE_STANDARD:
-				/* Take the value from the update. */
-				WT_ERR(__rec_cell_build_val(session, r,
-				    upd->data, upd->size,
-				    start_ts, stop_ts, start_txn, stop_txn, 0));
-				dictionary = true;
-				break;
-			case WT_UPDATE_TOMBSTONE:
-				/*
-				 * If this key/value pair was deleted, we're
-				 * done.
-				 *
-				 * Overflow keys referencing discarded values
-				 * are no longer useful, discard the backing
-				 * blocks.  Don't worry about reuse, reusing
-				 * keys from a row-store page reconciliation
-				 * seems unlikely enough to ignore.
-				 */
-				if (kpack != NULL && kpack->ovfl &&
-				    kpack->raw != WT_CELL_KEY_OVFL_RM) {
-					/*
-					 * Keys are part of the name-space, we
-					 * can't remove them from the in-memory
-					 * tree; if an overflow key was deleted
-					 * without being instantiated (for
-					 * example, cursor-based truncation), do
-					 * it now.
-					 */
-					if (ikey == NULL)
-						WT_ERR(__wt_row_leaf_key(
-						    session,
-						    page, rip, tmpkey, true));
-
-					WT_ERR(__wt_ovfl_discard_add(
-					    session, page, kpack->cell));
-				}
-
-				/*
-				 * We aren't actually creating the key so we
-				 * can't use bytes from this key to provide
-				 * prefix information for a subsequent key.
-				 */
-				tmpkey->size = 0;
-
-				/* Proceed with appended key/value pairs. */
-				goto leaf_insert;
-			WT_ILLEGAL_VALUE_ERR(session, upd->type);
-			}
-		}
-
-		/*
-		 * Build key cell.
-		 *
-		 * If the key is an overflow key that hasn't been removed, use
-		 * the original backing blocks.
-		 */
-		key_onpage_ovfl = kpack != NULL &&
-		    kpack->ovfl && kpack->raw != WT_CELL_KEY_OVFL_RM;
-		if (key_onpage_ovfl) {
-			key->buf.data = cell;
-			key->buf.size = __wt_cell_total_len(kpack);
-			key->cell_len = 0;
-			key->len = key->buf.size;
-			ovfl_key = true;
-
-			/*
-			 * We aren't creating a key so we can't use this key as
-			 * a prefix for a subsequent key.
-			 */
-			tmpkey->size = 0;
-
-			/* Track if page has overflow items. */
-			r->ovfl_items = true;
-		} else {
-			/*
-			 * Get the key from the page or an instantiated key, or
-			 * inline building the key from a previous key (it's a
-			 * fast path for simple, prefix-compressed keys), or by
-			 * by building the key from scratch.
-			 */
-			if (__wt_row_leaf_key_info(page, copy,
-			    NULL, &cell, &tmpkey->data, &tmpkey->size))
-				goto build;
-
-			kpack = &_kpack;
-			__wt_cell_unpack(session, page, cell, kpack);
-			if (btree->huffman_key == NULL &&
-			    kpack->type == WT_CELL_KEY &&
-			    tmpkey->size >= kpack->prefix) {
-				/*
-				 * The previous clause checked for a prefix of
-				 * zero, which means the temporary buffer must
-				 * have a non-zero size, and it references a
-				 * valid key.
-				 */
-				WT_ASSERT(session, tmpkey->size != 0);
-
-				/*
-				 * Grow the buffer as necessary, ensuring data
-				 * data has been copied into local buffer space,
-				 * then append the suffix to the prefix already
-				 * in the buffer.
-				 *
-				 * Don't grow the buffer unnecessarily or copy
-				 * data we don't need, truncate the item's data
-				 * length to the prefix bytes.
-				 */
-				tmpkey->size = kpack->prefix;
-				WT_ERR(__wt_buf_grow(session,
-				    tmpkey, tmpkey->size + kpack->size));
-				memcpy((uint8_t *)tmpkey->mem + tmpkey->size,
-				    kpack->data, kpack->size);
-				tmpkey->size += kpack->size;
-			} else
-				WT_ERR(__wt_row_leaf_key_copy(
-				    session, page, rip, tmpkey));
-build:
-			WT_ERR(__rec_cell_build_leaf_key(session, r,
-			    tmpkey->data, tmpkey->size, &ovfl_key));
-		}
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len)) {
-			/*
-			 * If we copied address blocks from the page rather than
-			 * building the actual key, we have to build the key now
-			 * because we are about to promote it.
-			 */
-			if (key_onpage_ovfl) {
-				WT_ERR(__wt_dsk_cell_data_ref(session,
-				    WT_PAGE_ROW_LEAF, kpack, r->cur));
-				WT_NOT_READ(key_onpage_ovfl, false);
-			}
-
-			/*
-			 * Turn off prefix compression until a full key written
-			 * to the new page, and (unless already working with an
-			 * overflow key), rebuild the key without compression.
-			 */
-			if (r->key_pfx_compress_conf) {
-				r->key_pfx_compress = false;
-				if (!ovfl_key)
-					WT_ERR(__rec_cell_build_leaf_key(
-					    session, r, NULL, 0, &ovfl_key));
-			}
-
-			WT_ERR(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-		}
-
-		/* Copy the key/value pair onto the page. */
-		__rec_image_copy(session, r, key);
-		if (val->len == 0 && __rec_row_zero_len(
-			session, start_ts, stop_ts, start_txn, stop_txn))
-			r->any_empty_value = true;
-		else {
-			r->all_empty_value = false;
-			if (dictionary && btree->dictionary)
-				WT_ERR(__rec_dict_replace(session, r,
-				    start_ts, stop_ts, start_txn, stop_txn,
-				    0, val));
-			__rec_image_copy(session, r, val);
-		}
-		__rec_addr_ts_update(r,
-		    start_ts, durable_ts, stop_ts, start_txn, stop_txn);
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-
-leaf_insert:	/* Write any K/V pairs inserted into the page after this key. */
-		if ((ins = WT_SKIP_FIRST(WT_ROW_INSERT(page, rip))) != NULL)
-			WT_ERR(__rec_row_leaf_insert(session, r, ins));
-	}
-
-	/* Write the remnant page. */
-	ret = __rec_split_finish(session, r);
-
-err:	__wt_scr_free(session, &tmpkey);
-	__wt_scr_free(session, &tmpval);
-	return (ret);
-}
-
-/*
- * __rec_row_leaf_insert --
- *	Walk an insert chain, writing K/V pairs.
- */
-static int
-__rec_row_leaf_insert(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_INSERT *ins)
-{
-	WT_BTREE *btree;
-	WT_CURSOR_BTREE *cbt;
-	WT_REC_KV *key, *val;
-	WT_UPDATE *upd;
-	WT_UPDATE_SELECT upd_select;
-	wt_timestamp_t start_ts, durable_ts, stop_ts;
-	uint64_t start_txn, stop_txn;
-	bool ovfl_key, upd_saved;
-
-	btree = S2BT(session);
-	cbt = &r->update_modify_cbt;
-
-	key = &r->k;
-	val = &r->v;
-
-	for (; ins != NULL; ins = WT_SKIP_NEXT(ins)) {
-		WT_RET(__wt_rec_upd_select(
-		    session, r, ins, NULL, NULL, &upd_select));
-		upd = upd_select.upd;
-		start_ts = upd_select.start_ts;
-		durable_ts = upd_select.durable_ts;
-		stop_ts = upd_select.stop_ts;
-		start_txn = upd_select.start_txn;
-		stop_txn = upd_select.stop_txn;
-		upd_saved = upd_select.upd_saved;
-
-		if (upd == NULL) {
-			/*
-			 * If no update is visible but some were saved, check
-			 * for splits.
-			 */
-			if (!upd_saved)
-				continue;
-			if (!__rec_need_split(r, WT_INSERT_KEY_SIZE(ins)))
-				continue;
-
-			/* Copy the current key into place and then split. */
-			WT_RET(__wt_buf_set(session, r->cur,
-			    WT_INSERT_KEY(ins), WT_INSERT_KEY_SIZE(ins)));
-			WT_RET(__rec_split_crossing_bnd(
-			    session, r, WT_INSERT_KEY_SIZE(ins)));
-
-			/*
-			 * Turn off prefix and suffix compression until a full
-			 * key is written into the new page.
-			 */
-			r->key_pfx_compress = r->key_sfx_compress = false;
-			continue;
-		}
-
-		switch (upd->type) {
-		case WT_UPDATE_MODIFY:
-			/*
-			 * Impossible slot, there's no backing on-page
-			 * item.
-			 */
-			cbt->slot = UINT32_MAX;
-			WT_RET(__wt_value_return_upd(
-			    session, cbt, upd, F_ISSET(r, WT_REC_VISIBLE_ALL)));
-			WT_RET(__rec_cell_build_val(session, r,
-			    cbt->iface.value.data, cbt->iface.value.size,
-			    start_ts, stop_ts, start_txn, stop_txn, 0));
-			break;
-		case WT_UPDATE_STANDARD:
-			/* Take the value from the update. */
-			WT_RET(__rec_cell_build_val(session, r,
-			    upd->data, upd->size,
-			    start_ts, stop_ts, start_txn, stop_txn, 0));
-			break;
-		case WT_UPDATE_TOMBSTONE:
-			continue;
-		WT_ILLEGAL_VALUE(session, upd->type);
-		}
-
-		/* Build key cell. */
-		WT_RET(__rec_cell_build_leaf_key(session, r,
-		    WT_INSERT_KEY(ins), WT_INSERT_KEY_SIZE(ins), &ovfl_key));
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len)) {
-			/*
-			 * Turn off prefix compression until a full key written
-			 * to the new page, and (unless already working with an
-			 * overflow key), rebuild the key without compression.
-			 */
-			if (r->key_pfx_compress_conf) {
-				r->key_pfx_compress = false;
-				if (!ovfl_key)
-					WT_RET(__rec_cell_build_leaf_key(
-					    session, r, NULL, 0, &ovfl_key));
-			}
-
-			WT_RET(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-		}
-
-		/* Copy the key/value pair onto the page. */
-		__rec_image_copy(session, r, key);
-		if (val->len == 0 && __rec_row_zero_len(
-		    session, start_ts, stop_ts, start_txn, stop_txn))
-			r->any_empty_value = true;
-		else {
-			r->all_empty_value = false;
-			if (btree->dictionary)
-				WT_RET(__rec_dict_replace(session, r,
-				    start_ts, stop_ts, start_txn, stop_txn,
-				    0, val));
-			__rec_image_copy(session, r, val);
-		}
-		__rec_addr_ts_update(r,
-		    start_ts, durable_ts, stop_ts, start_txn, stop_txn);
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-	}
 
 	return (0);
 }
@@ -5368,257 +2580,11 @@ __rec_las_wrapup_err(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 }
 
 /*
- * __rec_cell_build_int_key --
- *	Process a key and return a WT_CELL structure and byte string to be
- *	stored on a row-store internal page.
- */
-static int
-__rec_cell_build_int_key(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, const void *data, size_t size, bool *is_ovflp)
-{
-	WT_BTREE *btree;
-	WT_REC_KV *key;
-
-	*is_ovflp = false;
-
-	btree = S2BT(session);
-	key = &r->k;
-
-	/* Copy the bytes into the "current" and key buffers. */
-	WT_RET(__wt_buf_set(session, r->cur, data, size));
-	WT_RET(__wt_buf_set(session, &key->buf, data, size));
-
-	/* Create an overflow object if the data won't fit. */
-	if (size > btree->maxintlkey) {
-		WT_STAT_DATA_INCR(session, rec_overflow_key_internal);
-
-		*is_ovflp = true;
-		return (__rec_cell_build_ovfl(session, r, key, WT_CELL_KEY_OVFL,
-		    WT_TS_NONE, WT_TS_NONE, WT_TXN_NONE, WT_TXN_NONE, 0));
-	}
-
-	key->cell_len = __wt_cell_pack_int_key(&key->cell, key->buf.size);
-	key->len = key->cell_len + key->buf.size;
-
-	return (0);
-}
-
-/*
- * __rec_cell_build_leaf_key --
- *	Process a key and return a WT_CELL structure and byte string to be
- *	stored on a row-store leaf page.
- */
-static int
-__rec_cell_build_leaf_key(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, const void *data, size_t size, bool *is_ovflp)
-{
-	WT_BTREE *btree;
-	WT_REC_KV *key;
-	size_t pfx_max;
-	const uint8_t *a, *b;
-	uint8_t pfx;
-
-	*is_ovflp = false;
-
-	btree = S2BT(session);
-	key = &r->k;
-
-	pfx = 0;
-	if (data == NULL)
-		/*
-		 * When data is NULL, our caller has a prefix compressed key
-		 * they can't use (probably because they just crossed a split
-		 * point).  Use the full key saved when last called, instead.
-		 */
-		WT_RET(__wt_buf_set(
-		    session, &key->buf, r->cur->data, r->cur->size));
-	else {
-		/*
-		 * Save a copy of the key for later reference: we use the full
-		 * key for prefix-compression comparisons, and if we are, for
-		 * any reason, unable to use the compressed key we generate.
-		 */
-		WT_RET(__wt_buf_set(session, r->cur, data, size));
-
-		/*
-		 * Do prefix compression on the key.  We know by definition the
-		 * previous key sorts before the current key, which means the
-		 * keys must differ and we just need to compare up to the
-		 * shorter of the two keys.
-		 */
-		if (r->key_pfx_compress) {
-			/*
-			 * We can't compress out more than 256 bytes, limit the
-			 * comparison to that.
-			 */
-			pfx_max = UINT8_MAX;
-			if (size < pfx_max)
-				pfx_max = size;
-			if (r->last->size < pfx_max)
-				pfx_max = r->last->size;
-			for (a = data, b = r->last->data; pfx < pfx_max; ++pfx)
-				if (*a++ != *b++)
-					break;
-
-			/*
-			 * Prefix compression may cost us CPU and memory when
-			 * the page is re-loaded, don't do it unless there's
-			 * reasonable gain.
-			 */
-			if (pfx < btree->prefix_compression_min)
-				pfx = 0;
-			else
-				WT_STAT_DATA_INCRV(
-				    session, rec_prefix_compression, pfx);
-		}
-
-		/* Copy the non-prefix bytes into the key buffer. */
-		WT_RET(__wt_buf_set(
-		    session, &key->buf, (uint8_t *)data + pfx, size - pfx));
-	}
-
-	/* Optionally compress the key using the Huffman engine. */
-	if (btree->huffman_key != NULL)
-		WT_RET(__wt_huffman_encode(session, btree->huffman_key,
-		    key->buf.data, (uint32_t)key->buf.size, &key->buf));
-
-	/* Create an overflow object if the data won't fit. */
-	if (key->buf.size > btree->maxleafkey) {
-		/*
-		 * Overflow objects aren't prefix compressed -- rebuild any
-		 * object that was prefix compressed.
-		 */
-		if (pfx == 0) {
-			WT_STAT_DATA_INCR(session, rec_overflow_key_leaf);
-
-			*is_ovflp = true;
-			return (__rec_cell_build_ovfl(session, r, key,
-			    WT_CELL_KEY_OVFL, WT_TS_NONE, WT_TS_NONE,
-			    WT_TXN_NONE, WT_TXN_NONE, 0));
-		}
-		return (
-		    __rec_cell_build_leaf_key(session, r, NULL, 0, is_ovflp));
-	}
-
-	key->cell_len = __wt_cell_pack_leaf_key(&key->cell, pfx, key->buf.size);
-	key->len = key->cell_len + key->buf.size;
-
-	return (0);
-}
-
-/*
- * __rec_cell_build_addr --
- *	Process an address reference and return a cell structure to be stored
- *	on the page.
- */
-static void
-__rec_cell_build_addr(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_ADDR *addr, bool proxy_cell, uint64_t recno)
-{
-	WT_REC_KV *val;
-	u_int cell_type;
-
-	val = &r->v;
-
-	/*
-	 * Our caller optionally specifies a cell type (deleted proxy cells),
-	 * otherwise go with what we know.
-	 */
-	if (proxy_cell)
-		cell_type = WT_CELL_ADDR_DEL;
-	else {
-		switch (addr->type) {
-		case WT_ADDR_INT:
-			cell_type = WT_CELL_ADDR_INT;
-			break;
-		case WT_ADDR_LEAF:
-			cell_type = WT_CELL_ADDR_LEAF;
-			break;
-		case WT_ADDR_LEAF_NO:
-		default:
-			cell_type = WT_CELL_ADDR_LEAF_NO;
-			break;
-		}
-		WT_ASSERT(session, addr->size != 0);
-	}
-
-	/*
-	 * We don't check the address size because we can't store an address on
-	 * an overflow page: if the address won't fit, the overflow page's
-	 * address won't fit either.  This possibility must be handled by Btree
-	 * configuration, we have to disallow internal page sizes that are too
-	 * small with respect to the largest address cookie the underlying block
-	 * manager might return.
-	 */
-
-	/*
-	 * We don't copy the data into the buffer, it's not necessary; just
-	 * re-point the buffer's data/length fields.
-	 */
-	val->buf.data = addr->addr;
-	val->buf.size = addr->size;
-	val->cell_len = __wt_cell_pack_addr(
-	    session, &val->cell, cell_type, recno, addr->oldest_start_ts,
-	    addr->newest_durable_ts, addr->newest_stop_ts,
-	    addr->oldest_start_txn, addr->newest_stop_txn, val->buf.size);
-	val->len = val->cell_len + val->buf.size;
-}
-
-/*
- * __rec_cell_build_val --
- *	Process a data item and return a WT_CELL structure and byte string to
- *	be stored on the page.
- */
-static int
-__rec_cell_build_val(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    const void *data, size_t size,
-    wt_timestamp_t start_ts, wt_timestamp_t stop_ts,
-    uint64_t start_txn, uint64_t stop_txn, uint64_t rle)
-{
-	WT_BTREE *btree;
-	WT_REC_KV *val;
-
-	btree = S2BT(session);
-
-	val = &r->v;
-
-	/*
-	 * We don't copy the data into the buffer, it's not necessary; just
-	 * re-point the buffer's data/length fields.
-	 */
-	val->buf.data = data;
-	val->buf.size = size;
-
-	/* Handle zero-length cells quickly. */
-	if (size != 0) {
-		/* Optionally compress the data using the Huffman engine. */
-		if (btree->huffman_value != NULL)
-			WT_RET(__wt_huffman_encode(
-			    session, btree->huffman_value,
-			    val->buf.data, (uint32_t)val->buf.size, &val->buf));
-
-		/* Create an overflow object if the data won't fit. */
-		if (val->buf.size > btree->maxleafvalue) {
-			WT_STAT_DATA_INCR(session, rec_overflow_value);
-
-			return (__rec_cell_build_ovfl(session, r, val,
-			    WT_CELL_VALUE_OVFL,
-			    start_ts, stop_ts, start_txn, stop_txn, rle));
-		}
-	}
-	val->cell_len = __wt_cell_pack_value(session, &val->cell,
-	    start_ts, stop_ts, start_txn, stop_txn, rle, val->buf.size);
-	val->len = val->cell_len + val->buf.size;
-
-	return (0);
-}
-
-/*
- * __rec_cell_build_ovfl --
+ * __wt_rec_cell_build_ovfl --
  *	Store overflow items in the file, returning the address cookie.
  */
-static int
-__rec_cell_build_ovfl(WT_SESSION_IMPL *session,
+int
+__wt_rec_cell_build_ovfl(WT_SESSION_IMPL *session,
     WT_RECONCILE *r, WT_REC_KV *kv, uint8_t type,
     wt_timestamp_t start_ts, wt_timestamp_t stop_ts,
     uint64_t start_txn, uint64_t stop_txn, uint64_t rle)
